Refactor generator imports and update test cases
- Updated imports in the pipeline and test files to use the new generator module.
- Replaced instances of `generator_v2` with `generator` across various test files and integration tests.
- Added temporary directory fixtures in golden test files to enhance test isolation.
- Adjusted assertions in tests to accommodate changes in the generator's output structure.
- Ensured that all references to the old generator module are removed for consistency.
