+++ docs/plan-2025-09-13-01.md
@@ -3,8 +3,8 @@ Documenting the generator overhaul plan and acceptance criteria.
 Overview
 --------
 This plan describes a clean, testable rewrite of the generator stage used by
-the unittest->pytest converter. The existing `stages/generator.py` contains
-many heuristics and ad-hoc rewrites; `generator_v2` will provide a smaller,
+ the unittest->pytest converter. The existing `stages/generator.py` contains
+ many heuristics and ad-hoc rewrites; `generator` will provide a smaller,
 well-documented core and clear helpers that produce the same public output
 shape used by downstream stages (specs and fixture_nodes).
 
@@ -41,16 +41,16 @@ Plan
 - Implement small helpers: build_local_assignments, rewrite_cleanup,
   emit_per_attr_fixture_node. The implementation will intentionally avoid any
   composite/grouping helpers.
-- Add unit tests for each helper and integration-style tests that invoke
-  `generator_v2` on representative CollectorOutput instances and check
+ - Add unit tests for each helper and integration-style tests that invoke
+  `generator` on representative CollectorOutput instances and check
   high-level properties (one fixture per attribute, mkdtemp preserved,
   cleanup attached to the correct fixture).
 - Iterate against `tests/unit/test_converter_goldens_strict.py` until green.
 
 Notes
 -----
-This plan intentionally keeps the v2 generator isolated (module suffix _v2)
-so downstream stages can be migrated incrementally.
+ This plan documents the generator implementation (module `stages/generator`)
+ so downstream stages can be migrated incrementally.
 # Converter Tweak Plan — Unittest → Pytest (2025-09-13)
 
 ## Summary
@@ -144,13 +144,16 @@ Scope: changes limited to the generator and converter stages, import injector, a
 
 ```python
 @pytest.fixture
-def temp_dirs() -> Generator[Dict[str, Path], None, None]:
-    temp_dir = Path(tempfile.mkdtemp())
-    try:
-        yield {"temp_dir": temp_dir}
-    finally:
-        shutil.rmtree(temp_dir, ignore_errors=True)
+def temp_dir() -> Path:
+  # create a temporary directory per-attribute and return it. If cleanup is
+  # required, the generator-style fixture with try/finally and yield is used.
+  return Path(tempfile.mkdtemp())
 ```
+
+Note: the generator emits per-attribute fixtures (e.g., `temp_dir`,
+`config_dir`) rather than a single composite `temp_dirs` mapping. When a
+fixture performs cleanup it will be emitted as a generator fixture using
+`yield` inside a `try/finally` block.
 
 - Simple direct-return fixture:
 
@@ -204,3 +207,7 @@ Document created 2025-09-13.
   - These tests pass locally in the development environment.
 - `import_injector` was updated to merge missing typing names into an existing `from typing import ...` import when present, and to insert a single `from pathlib import Path` when `Path` is requested.
 - Next recommended step: add golden-based tests for the converter output (samples → canonical `*-pytest.txt`) and run ruff/mypy across the repo before making further generator-level changes.
+
+## Post-implementation note (2025-09-13)
+
+- The legacy generator implementation `splurge_unittest_to_pytest/stages/generator.py` has been replaced by the current `generator` implementation. The repository and tests were validated after migration (ruff/mypy/pytest green at time of migration). The canonical behavior is: per-attribute fixtures (no composite `temp_dirs`) and accurate typing import collection for emitted annotations.
