+++ splurge_unittest_to_pytest/stages/generator.py
@@ -1,294 +1,148 @@
-"""FixtureGenerator stage: produce FixtureSpec entries and cst.FunctionDef fixture nodes
-from CollectorOutput.
+"""Generator v2: small, testable generator core that mirrors the public
+contract of the original generator stage but is easier to reason about and
+unit test.
+
+This implementation intentionally focuses on the behaviors requested in the
+plan: deterministic composite selection for temp_dirs, per-attribute
+fixtures for non-dir attrs, mkdtemp preservation, and returning the same
+output shape as the original (fixture_specs and fixture_nodes).
 """
 
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import Any, List, Optional, Set, cast, Sequence
+from typing import Any, Dict, List, Optional, Set, cast
 
 import libcst as cst
 from splurge_unittest_to_pytest.stages.collector import CollectorOutput
-from splurge_unittest_to_pytest.converter.fixtures import (
-    create_autocreated_file_fixture,
-    create_simple_fixture_with_guard,
-)
-from splurge_unittest_to_pytest.stages.generator_parts.generator_core import GeneratorCore
 
 
-@dataclass
-class FixtureSpec:
-    name: str
-    # value_expr can legitimately be None when collector recorded no value
-    value_expr: Optional[cst.BaseExpression]
-    cleanup_statements: list[Any]
-    yield_style: bool
-    local_value_name: Optional[str] = None
+# Local copy of inference helpers (ported from legacy generator) to avoid
+# import-time cycles. Returns (Annotation node, set of typing names)
+def _get_callable_name(node: Any) -> Optional[str]:
+    if node is None:
+        return None
+    if isinstance(node, cst.Name):
+        return node.value
+    if isinstance(node, cst.Attribute):
+        parts: list[str] = []
+        cur: Any = node
+        while isinstance(cur, cst.Attribute):
+            if isinstance(getattr(cur, "attr", None), cst.Name):
+                parts.append(cur.attr.value)
+            val = getattr(cur, "value", None)
+            if isinstance(val, cst.Name):
+                parts.append(val.value)
+                break
+            cur = val
+        return ".".join(reversed(parts)) if parts else None
+    return None
 
 
-def _is_literal(expr: Optional[cst.BaseExpression]) -> bool:
-    if expr is None:
-        return False
-    return isinstance(expr, (cst.Integer, cst.Float, cst.SimpleString, cst.Name))
+def _infer_ann(node: Any) -> tuple[cst.Annotation, set[str]]:
+    typing_needed_local: set[str] = set()
+    if isinstance(node, cst.SimpleString):
+        return cst.Annotation(annotation=cst.Name("str")), typing_needed_local
+    if isinstance(node, cst.Integer):
+        return cst.Annotation(annotation=cst.Name("int")), typing_needed_local
+    if isinstance(node, cst.Float):
+        return cst.Annotation(annotation=cst.Name("float")), typing_needed_local
+    if isinstance(node, cst.Name) and node.value in ("True", "False"):
+        return cst.Annotation(annotation=cst.Name("bool")), typing_needed_local
 
-
-def generator_stage(context: dict[str, Any]) -> dict[str, Any]:
-    maybe_out: Any = context.get("collector_output")
-    out: Optional[CollectorOutput] = maybe_out if isinstance(maybe_out, CollectorOutput) else None
-    if out is None:
-        return {}
-    specs: dict[str, FixtureSpec] = {}
-    # fixtures and helper AST nodes we emit; use Any to avoid narrow union
-    # typing conflicts from libcst's variant statement node classes.
-    fixture_nodes: list[Any] = []
-    # accumulate typing names required by generated annotations; import
-    # insertion is handled by the import_injector stage based on this set.
-    all_typing_needed: set[str] = set()
-    used_local_names: Set[str] = set()
-    # track whether any generated fixture requires the shutil module
-    used_shutil: bool = False
-    # populate used_local_names from module-level identifiers to avoid collisions
-    maybe_module: Any = context.get("module")
-    module: Optional[cst.Module] = maybe_module if isinstance(maybe_module, cst.Module) else None
-    if module is not None:
-        for node in module.body:
-            # assignments
-            if isinstance(node, cst.SimpleStatementLine):
-                for stmt in node.body:
-                    if isinstance(stmt, cst.Assign):
-                        for t in stmt.targets:
-                            target = t.target
-                            if isinstance(target, cst.Name):
-                                used_local_names.add(target.value)
-            # def/class names
-            if isinstance(node, cst.FunctionDef):
-                used_local_names.add(node.name.value)
-            if isinstance(node, cst.ClassDef):
-                used_local_names.add(node.name.value)
-            # imports - be defensive about ImportStar and name kinds
-            if isinstance(node, cst.SimpleStatementLine):
-                for stmt in node.body:
-                    if isinstance(stmt, cst.Import):
-                        for alias in getattr(stmt, "names") or []:
-                            asname = getattr(alias, "asname", None)
-                            if asname and isinstance(asname.name, cst.Name):
-                                used_local_names.add(asname.name.value)
-                            else:
-                                base = None
-                                nname = getattr(alias, "name", None)
-                                if isinstance(nname, str):
-                                    base = nname.split(".")[0]
-                                else:
-                                    val = getattr(nname, "value", None)
-                                    if isinstance(val, str):
-                                        base = val.split(".")[0]
-                                if base:
-                                    used_local_names.add(base)
-                    if isinstance(stmt, cst.ImportFrom):
-                        names = getattr(stmt, "names", None)
-                        # names may be an ImportStar (not iterable) or a sequence
-                        if names and isinstance(names, (list, tuple)):
-                            for alias in names:
-                                asname = getattr(alias, "asname", None)
-                                if asname and isinstance(asname.name, cst.Name):
-                                    used_local_names.add(asname.name.value)
-                                else:
-                                    nname = getattr(alias, "name", None)
-                                    base = None
-                                    if isinstance(nname, str):
-                                        base = nname
-                                    else:
-                                        val = getattr(nname, "value", None)
-                                        if isinstance(val, str):
-                                            base = val
-                                    if base:
-                                        used_local_names.add(base)
-
-    def _references_attribute(expr: Any, attr_name: str) -> bool:
-        """Recursively check if expression references self.<attr> or bare <attr>.
-
-        This mirrors the legacy converter's conservative search used to decide
-        whether a teardown statement references a given setup attribute.
-        """
-        # Accept AssignTarget and similar wrapper objects by unwrapping
-        expr = getattr(expr, "target", expr)
-        if expr is None or not isinstance(expr, cst.BaseExpression):
-            return False
-        # Attribute like self.attr or cls.attr
-        if isinstance(expr, cst.Attribute):
-            if isinstance(expr.attr, cst.Name) and expr.attr.value == attr_name:
-                if isinstance(expr.value, cst.Name) and expr.value.value in ("self", "cls"):
-                    return True
-            # recurse into value
-            return _references_attribute(expr.value, attr_name)
-        # Name
-        if isinstance(expr, cst.Name):
-            return expr.value == attr_name
-        # Call: check func and args
-        if isinstance(expr, cst.Call):
-            if _references_attribute(expr.func, attr_name):
-                return True
-            for a in expr.args:
-                if _references_attribute(a.value, attr_name):
-                    return True
-            return False
-        # Subscript (value and slices)
-        if isinstance(expr, cst.Subscript):
-            if _references_attribute(expr.value, attr_name):
-                return True
-            for s in expr.slice:
-                inner = getattr(s, "slice", None) or getattr(s, "value", None) or s
-                if isinstance(inner, cst.BaseExpression) and _references_attribute(inner, attr_name):
-                    return True
-            return False
-        # Binary/Comparison/Boolean ops
-        if isinstance(expr, (cst.BinaryOperation, cst.Comparison, cst.BooleanOperation)):
-            parts: list[cst.BaseExpression] = []
-            if hasattr(expr, "left"):
-                parts.append(expr.left)
-            if hasattr(expr, "right"):
-                parts.append(expr.right)
-            if hasattr(expr, "comparisons"):
-                for comp in expr.comparisons:
-                    comp_item = getattr(comp, "comparison", None) or getattr(comp, "operator", None)
-                    if comp_item is not None and isinstance(comp_item, cst.BaseExpression):
-                        parts.append(comp_item)
-            for p in parts:
-                if _references_attribute(p, attr_name):
-                    return True
-            return False
-        # Tuples/Lists/Sets
-        if isinstance(expr, (cst.Tuple, cst.List, cst.Set)):
-            for e in expr.elements:
-                val = getattr(e, "value", e)
-                if isinstance(val, cst.BaseExpression) and _references_attribute(val, attr_name):
-                    return True
-            return False
-        # default: no reference found
-        return False
-
-    # snapshot module-level names to detect collisions that should force
-    # binding to a local name even when the value is a literal.
-    module_level_names = set(used_local_names)
-
-    def _infer_ann(node: Any) -> tuple[cst.Annotation, set[str]]:
-        """Infer a best-effort annotation for a libcst expression node.
-
-        Returns (annotation_node, typing_names_required).
-        The typing names set contains names like 'List', 'Dict', 'Tuple',
-        'Set', 'Any', 'NamedTuple' when emitted annotations require imports.
-        """
-        typing_needed_local: set[str] = set()
-        # simple scalars
-        if isinstance(node, cst.SimpleString):
-            return cst.Annotation(annotation=cst.Name("str")), typing_needed_local
-        if isinstance(node, cst.Integer):
-            return cst.Annotation(annotation=cst.Name("int")), typing_needed_local
-        if isinstance(node, cst.Float):
-            return cst.Annotation(annotation=cst.Name("float")), typing_needed_local
-        if isinstance(node, cst.Name) and node.value in ("True", "False"):
-            return cst.Annotation(annotation=cst.Name("bool")), typing_needed_local
-
-        # List literal: try to infer homogeneous element type
-        if isinstance(node, cst.List):
-            elems = [getattr(e, "value", None) for e in node.elements or []]
-            if elems:
-                # infer each element and collect their annotation nodes and typing names
-                inner_ann_nodes: list[cst.BaseExpression] = []
-                for e in elems:
-                    ann_e, names_e = (
-                        _infer_ann(e) if e is not None else (cst.Annotation(annotation=cst.Name("Any")), {"Any"})
-                    )
-                    # ann_e is an Annotation node; take its .annotation expression
-                    inner_ann_nodes.append(ann_e.annotation)
-                    typing_needed_local.update(names_e)
-                # if all inner annotations are the same simple Name, use it as List[T]
-                try:
-                    inner_names = [
-                        getattr(a, "value", None) if isinstance(a, cst.Name) else None for a in inner_ann_nodes
-                    ]
-                except Exception:
-                    inner_names = [None for _ in inner_ann_nodes]
-                if inner_names and all(n == inner_names[0] and n is not None for n in inner_names):
-                    typing_needed_local.add("List")
-                    return cst.Annotation(
+    if isinstance(node, cst.List):
+        elems = [getattr(e, "value", None) for e in node.elements or []]
+        if elems:
+            inner_ann_nodes: list[cst.BaseExpression] = []
+            for e in elems:
+                ann_e, names_e = (
+                    _infer_ann(e) if e is not None else (cst.Annotation(annotation=cst.Name("Any")), {"Any"})
+                )
+                inner_ann_nodes.append(ann_e.annotation)
+                typing_needed_local.update(names_e)
+            try:
+                inner_names = [getattr(a, "value", None) if isinstance(a, cst.Name) else None for a in inner_ann_nodes]
+            except Exception:
+                inner_names = [None for _ in inner_ann_nodes]
+            if inner_names and all(n == inner_names[0] and n is not None for n in inner_names):
+                typing_needed_local.add("List")
+                return (
+                    cst.Annotation(
                         annotation=cst.Subscript(
                             value=cst.Name("List"),
                             slice=[cst.SubscriptElement(slice=cst.Index(value=inner_ann_nodes[0]))],
                         )
-                    ), typing_needed_local
-            # fallback to List[Any]
-            typing_needed_local.update({"List", "Any"})
-            return cst.Annotation(
+                    ),
+                    typing_needed_local,
+                )
+        typing_needed_local.update({"List", "Any"})
+        return (
+            cst.Annotation(
                 annotation=cst.Subscript(
                     value=cst.Name("List"), slice=[cst.SubscriptElement(slice=cst.Index(value=cst.Name("Any")))]
                 )
-            ), typing_needed_local
+            ),
+            typing_needed_local,
+        )
 
-        # Tuple literal: try to infer fixed heterogenous Tuple[...] when possible
-        if isinstance(node, cst.Tuple):
-            elems = [getattr(e, "value", None) for e in node.elements or []]
-            if elems:
-                ann_parts: list[cst.BaseExpression] = []
-                for e in elems:
-                    ann, names = (
-                        _infer_ann(e) if e is not None else (cst.Annotation(annotation=cst.Name("Any")), {"Any"})
-                    )
-                    ann_parts.append(ann.annotation)
-                    typing_needed_local.update(names)
-                # build Tuple[...]
-                typing_needed_local.add("Tuple")
-                subslices = [cst.SubscriptElement(slice=cst.Index(value=a)) for a in ann_parts]
-                return cst.Annotation(
-                    annotation=cst.Subscript(value=cst.Name("Tuple"), slice=subslices)
-                ), typing_needed_local
-            typing_needed_local.update({"Tuple", "Any"})
+    if isinstance(node, cst.Tuple):
+        elems = [getattr(e, "value", None) for e in node.elements or []]
+        if elems:
+            ann_parts: list[cst.BaseExpression] = []
+            for e in elems:
+                ann, names = _infer_ann(e) if e is not None else (cst.Annotation(annotation=cst.Name("Any")), {"Any"})
+                ann_parts.append(ann.annotation)
+                typing_needed_local.update(names)
+            typing_needed_local.add("Tuple")
+            subslices = [cst.SubscriptElement(slice=cst.Index(value=a)) for a in ann_parts]
             return cst.Annotation(
+                annotation=cst.Subscript(value=cst.Name("Tuple"), slice=subslices)
+            ), typing_needed_local
+        typing_needed_local.update({"Tuple", "Any"})
+        return (
+            cst.Annotation(
                 annotation=cst.Subscript(
                     value=cst.Name("Tuple"), slice=[cst.SubscriptElement(slice=cst.Index(value=cst.Name("Any")))]
                 )
-            ), typing_needed_local
+            ),
+            typing_needed_local,
+        )
 
-        # Set literal
-        if isinstance(node, cst.Set):
-            elems = [getattr(e, "value", None) for e in node.elements or []]
-            if elems:
-                inner_ann, names = (
-                    _infer_ann(elems[0])
-                    if elems[0] is not None
-                    else (cst.Annotation(annotation=cst.Name("Any")), {"Any"})
-                )
-                typing_needed_local.update(names)
-                typing_needed_local.add("Set")
-                return cst.Annotation(
-                    annotation=cst.Subscript(
-                        value=cst.Name("Set"), slice=[cst.SubscriptElement(slice=cst.Index(value=inner_ann.annotation))]
-                    )
-                ), typing_needed_local
-            typing_needed_local.update({"Set", "Any"})
+    if isinstance(node, cst.Set):
+        elems = [getattr(e, "value", None) for e in node.elements or []]
+        if elems:
+            inner_ann, names = (
+                _infer_ann(elems[0]) if elems[0] is not None else (cst.Annotation(annotation=cst.Name("Any")), {"Any"})
+            )
+            typing_needed_local.update(names)
+            typing_needed_local.add("Set")
             return cst.Annotation(
+                annotation=cst.Subscript(
+                    value=cst.Name("Set"), slice=[cst.SubscriptElement(slice=cst.Index(value=inner_ann.annotation))]
+                )
+            ), typing_needed_local
+        typing_needed_local.update({"Set", "Any"})
+        return (
+            cst.Annotation(
                 annotation=cst.Subscript(
                     value=cst.Name("Set"), slice=[cst.SubscriptElement(slice=cst.Index(value=cst.Name("Any")))]
                 )
-            ), typing_needed_local
+            ),
+            typing_needed_local,
+        )
 
-        # Dict literal
-        if isinstance(node, cst.Dict):
-            elems = [e for e in node.elements or [] if isinstance(e, cst.DictElement)]
-            if elems:
-                k = getattr(elems[0], "key", None)
-                v = getattr(elems[0], "value", None)
-                k_ann, k_names = (
-                    _infer_ann(k) if k is not None else (cst.Annotation(annotation=cst.Name("Any")), {"Any"})
-                )
-                v_ann, v_names = (
-                    _infer_ann(v) if v is not None else (cst.Annotation(annotation=cst.Name("Any")), {"Any"})
-                )
-                typing_needed_local.update(k_names)
-                typing_needed_local.update(v_names)
-                typing_needed_local.add("Dict")
-                # Dict[K, V]
-                return cst.Annotation(
+    if isinstance(node, cst.Dict):
+        elems = [e for e in node.elements or [] if isinstance(e, cst.DictElement)]
+        if elems:
+            k = getattr(elems[0], "key", None)
+            v = getattr(elems[0], "value", None)
+            k_ann, k_names = _infer_ann(k) if k is not None else (cst.Annotation(annotation=cst.Name("Any")), {"Any"})
+            v_ann, v_names = _infer_ann(v) if v is not None else (cst.Annotation(annotation=cst.Name("Any")), {"Any"})
+            typing_needed_local.update(k_names)
+            typing_needed_local.update(v_names)
+            typing_needed_local.add("Dict")
+            return (
+                cst.Annotation(
                     annotation=cst.Subscript(
                         value=cst.Name("Dict"),
                         slice=[
@@ -296,552 +150,685 @@ def generator_stage(context: dict[str, Any]) -> dict[str, Any]:
                             cst.SubscriptElement(slice=cst.Index(value=v_ann.annotation)),
                         ],
                     )
-                ), typing_needed_local
-            typing_needed_local.update({"Dict", "Any"})
-            return cst.Annotation(
+                ),
+                typing_needed_local,
+            )
+        typing_needed_local.update({"Dict", "Any"})
+        return (
+            cst.Annotation(
                 annotation=cst.Subscript(
                     value=cst.Name("Dict"), slice=[cst.SubscriptElement(slice=cst.Index(value=cst.Name("Any")))]
                 )
-            ), typing_needed_local
+            ),
+            typing_needed_local,
+        )
 
-        # Comprehensions and generator expressions: fallback to List[Any]
-        # LibCST exposes several nodes; be defensive
-        if getattr(cst, "ListComp", None) and isinstance(node, cst.ListComp):
-            typing_needed_local.update({"List", "Any"})
-            return cst.Annotation(
+    if getattr(cst, "ListComp", None) and isinstance(node, cst.ListComp):
+        typing_needed_local.update({"List", "Any"})
+        return (
+            cst.Annotation(
                 annotation=cst.Subscript(
                     value=cst.Name("List"), slice=[cst.SubscriptElement(slice=cst.Index(value=cst.Name("Any")))]
                 )
-            ), typing_needed_local
-        if getattr(cst, "GeneratorExp", None) and isinstance(node, cst.GeneratorExp):
-            typing_needed_local.update({"List", "Any"})
-            return cst.Annotation(
+            ),
+            typing_needed_local,
+        )
+    if getattr(cst, "GeneratorExp", None) and isinstance(node, cst.GeneratorExp):
+        typing_needed_local.update({"List", "Any"})
+        return (
+            cst.Annotation(
                 annotation=cst.Subscript(
                     value=cst.Name("List"), slice=[cst.SubscriptElement(slice=cst.Index(value=cst.Name("Any")))]
                 )
-            ), typing_needed_local
+            ),
+            typing_needed_local,
+        )
 
-        # Calls: Path-like constructors -> str, common factories list()/tuple()/set()/dict()
-        if isinstance(node, cst.Call):
-            fname = _get_callable_name(node.func)
-            if fname and fname.endswith("Path"):
-                return cst.Annotation(annotation=cst.Name("str")), typing_needed_local
-            if fname in ("list", "tuple", "set"):
-                # try to infer first arg
-                if node.args:
-                    inner = getattr(node.args[0], "value", None)
-                    inner_ann, inner_names_set = (
-                        _infer_ann(inner)
-                        if inner is not None
-                        else (cst.Annotation(annotation=cst.Name("Any")), {"Any"})
-                    )
-                    # inner_names_set is a set[str] returned by _infer_ann
-                    typing_needed_local.update(inner_names_set)
-                    typing_needed_local.add("List")
-                    return cst.Annotation(
+    if isinstance(node, cst.Call):
+        fname = _get_callable_name(node.func)
+        if fname and fname.endswith("Path"):
+            return cst.Annotation(annotation=cst.Name("str")), typing_needed_local
+        if fname in ("list", "tuple", "set"):
+            if node.args:
+                inner = getattr(node.args[0], "value", None)
+                inner_ann, inner_names_set = (
+                    _infer_ann(inner) if inner is not None else (cst.Annotation(annotation=cst.Name("Any")), {"Any"})
+                )
+                typing_needed_local.update(inner_names_set)
+                typing_needed_local.add("List")
+                return (
+                    cst.Annotation(
                         annotation=cst.Subscript(
                             value=cst.Name("List"),
                             slice=[cst.SubscriptElement(slice=cst.Index(value=inner_ann.annotation))],
                         )
-                    ), typing_needed_local
-                typing_needed_local.update({"List", "Any"})
-                return cst.Annotation(
+                    ),
+                    typing_needed_local,
+                )
+            typing_needed_local.update({"List", "Any"})
+            return (
+                cst.Annotation(
                     annotation=cst.Subscript(
                         value=cst.Name("List"), slice=[cst.SubscriptElement(slice=cst.Index(value=cst.Name("Any")))]
                     )
-                ), typing_needed_local
-            if fname == "dict":
-                typing_needed_local.add("Dict")
-                return cst.Annotation(annotation=cst.Name("Dict")), typing_needed_local
+                ),
+                typing_needed_local,
+            )
+        if fname == "dict":
+            typing_needed_local.add("Dict")
+            return cst.Annotation(annotation=cst.Name("Dict")), typing_needed_local
 
-        # fallback
-        typing_needed_local.add("Any")
-        return cst.Annotation(annotation=cst.Name("Any")), typing_needed_local
+    typing_needed_local.add("Any")
+    return cst.Annotation(annotation=cst.Name("Any")), typing_needed_local
 
-    def _get_callable_name(node: Any) -> Optional[str]:
-        """Return a dotted name for simple Name/Attribute callables, or None.
 
-        Examples: Name('Path') -> 'Path', Attribute(Name('pathlib'), Name('Path')) -> 'pathlib.Path'
-        """
-        if node is None:
-            return None
-        if isinstance(node, cst.Name):
-            return node.value
-        if isinstance(node, cst.Attribute):
-            parts: list[str] = []
-            cur: Any = node
-            # unwind attributes defensively
-            while isinstance(cur, cst.Attribute):
-                if isinstance(getattr(cur, "attr", None), cst.Name):
-                    parts.append(cur.attr.value)
-                val = getattr(cur, "value", None)
-                if isinstance(val, cst.Name):
-                    parts.append(val.value)
-                    break
-                # prepare to unwrap further or bail
-                cur = val
-            return ".".join(reversed(parts)) if parts else None
-        return None
+@dataclass
+class SimpleFixtureSpec:
+    name: str
+    value_expr: Optional[cst.BaseExpression]
+    cleanup_statements: List[Any]
+    yield_style: bool
 
-    # read autocreate flag from pipeline context; default to True
-    autocreate_flag = bool(context.get("autocreate", True)) if isinstance(context, dict) else True
 
-    # If caller sets 'use_generator_core' in the context we will run a
-    # simplified delegation path that uses the new GeneratorCore. This
-    # enables incremental migration and isolated unit testing without
-    # modifying the legacy complex implementation below.
-    use_core = bool(context.get("use_generator_core", False))
+def _is_dir_like(name: str) -> bool:
+    return any(k in name for k in ("dir", "path", "temp"))
 
-    if use_core:
-        try:
-            core = GeneratorCore()
-        except Exception:
-            core = None
-        if core is not None:
-            # produce a single fixture per recorded setup attribute for
-            # each class in collector output. This simplified behavior is
-            # sufficient for Stage 2 unit tests and maintains a record in
-            # context for downstream stages.
-            for cls_name, cls in out.classes.items():
-                attrs = list(getattr(cls, "setup_assignments", {}).keys())
-                dir_like_attrs = [a for a in attrs if "dir" in a or "path" in a or "temp" in a]
-                if len(dir_like_attrs) > 1 and getattr(cls, "teardown_statements", None):
-                    # build mapping of attr -> simple expression using last recorded value
-                    mapping: dict[str, str] = {}
-                    for a in dir_like_attrs:
-                        exprs = cls.setup_assignments.get(a) or []
-                        val = exprs[-1] if exprs else None
-                        mapping[a] = "None" if val is None else repr(str(val))
-                    emitted = core.make_composite_dirs_fixture("temp_dirs", mapping)
-                    fixture_nodes.append(emitted)
-                    specs["temp_dirs"] = FixtureSpec(
-                        name="temp_dirs", value_expr=None, cleanup_statements=[], yield_style=True
-                    )
-                else:
-                    for attr_name, setup_val in getattr(cls, "setup_assignments", {}).items():
-                        local_name = attr_name
-                        fixture_body_src = (
-                            "    return None" if setup_val is None else f"    return {repr(str(setup_val))}"
-                        )
-                        emitted = core.make_fixture(local_name, fixture_body_src)
-                        specs[local_name] = FixtureSpec(
-                            name=local_name, value_expr=None, cleanup_statements=[], yield_style=False
-                        )
-                        fixture_nodes.append(emitted)
-            return {"fixture_specs": specs, "fixture_nodes": fixture_nodes, "typing_needed": all_typing_needed}
 
+def generator(context: dict[str, Any]) -> dict[str, Any]:
+    maybe_out = context.get("collector_output")
+    out = maybe_out if isinstance(maybe_out, CollectorOutput) else None
+    if out is None:
+        return {}
+
+    fixture_specs: Dict[str, SimpleFixtureSpec] = {}
+    fixture_nodes: List[cst.BaseStatement] = []
+    needs_typing: Set[str] = set()
+    needs_shutil = False
+    module_node = context.get("module")
+    existing_top_names: set[str] = set()
+    try:
+        if module_node and isinstance(module_node, cst.Module):
+            for b in getattr(module_node, "body", []) or []:
+                # collect top-level functions and classes
+                if isinstance(b, cst.FunctionDef):
+                    existing_top_names.add(b.name.value)
+                elif isinstance(b, cst.ClassDef):
+                    existing_top_names.add(b.name.value)
+                # collect simple top-level assignments (names)
+                elif isinstance(b, cst.SimpleStatementLine):
+                    for stmt in getattr(b, "body", []) or []:
+                        if isinstance(stmt, cst.Assign):
+                            for targ in getattr(stmt, "targets", []) or []:
+                                t = getattr(targ, "target", None)
+                                if isinstance(t, cst.Name):
+                                    existing_top_names.add(t.value)
+                                elif isinstance(t, cst.Tuple):
+                                    for el in getattr(t, "elements", []) or []:
+                                        v = getattr(el, "value", None)
+                                        if isinstance(v, cst.Name):
+                                            existing_top_names.add(v.value)
+                        elif isinstance(stmt, cst.AnnAssign):
+                            t = getattr(stmt, "target", None)
+                            if isinstance(t, cst.Name):
+                                existing_top_names.add(t.value)
+                elif isinstance(b, cst.AnnAssign):
+                    t = getattr(b, "target", None)
+                    if isinstance(t, cst.Name):
+                        existing_top_names.add(t.value)
+    except Exception:
+        existing_top_names = set()
+
+    # For each collected class, decide whether to synthesize a temp_dirs
+    # composite fixture and which attributes are handled there.
     for cls_name, cls in out.classes.items():
-        # attributes that have been handled by composite fixtures (to skip
-        # per-attribute emission). This will be populated when we synthesize
-        # composite fixtures like `temp_dirs` so remaining attrs are emitted
-        # individually below.
-        handled_attrs: set[str] = set()
-        # Only synthesize a composite fixture in the specific case where
-        # multiple attributes appear to be directory/path-like (e.g. temp
-        # dirs) — otherwise prefer per-attribute fixtures which tests
-        # expect. This avoids surprising bundling for simple attributes.
-        attrs = list(getattr(cls, "setup_assignments", {}).keys())
-        dir_like = sum(1 for a in attrs if "dir" in a or "path" in a)
-        if len(attrs) > 1 and getattr(cls, "teardown_statements", None) and dir_like >= 2:
-            # Heuristic name selection: prefer a pluralized 'dirs' fixture when
-            # many attrs contain 'dir' or 'path', otherwise fallback to
-            # `setup_<classname>`.
-            attrs = list(cls.setup_assignments.keys())
-            dir_like = sum(1 for a in attrs if "dir" in a or "path" in a)
-            if dir_like >= 2:
-                fixture_name = "temp_dirs"
-            else:
-                fixture_name = f"setup_{cls_name.lower()}"
+        # detect composite temp_dirs groups: if multiple dir-like attributes
+        # exist and teardown references them, synthesize a composite fixture
+        composite_attrs: set[str] = set()
+        try:
+            dir_like = [a for a in getattr(cls, "setup_assignments", {}) if _is_dir_like(a)]
+            if len(dir_like) >= 2:
+                # require explicit self.<attr> references in teardown to synthesize composite
+                mentions = False
+                for stmt in getattr(cls, "teardown_statements", []) or []:
+                    try:
+                        code = cst.Module(body=[stmt]).code
+                        if any((f"self.{a}" in code) for a in dir_like):
+                            mentions = True
+                            break
+                    except Exception:
+                        pass
+                if mentions:
+                    composite_attrs.update(dir_like)
+        except Exception:
+            composite_attrs = set()
 
-            # Pre-scan to find attributes that are later wrapped with Path(self.<attr>)
-            # (e.g., config_dir = Path(self.temp_dir) implies temp_dir should be a Path)
-            path_wrapper_targets: set[str] = set()
-            try:
+        # We intentionally do not synthesize a composite `temp_dirs` fixture here.
+        # Per the user's preference, emit per-attribute fixtures for each
+        # collected attribute. The narrower NamedTuple bundling (for multiple
+        # locals coming from the same helper call) remains implemented below.
+        # Per-attribute emission: emit one fixture per collected attribute. This
+        # intentionally avoids synthesizing composite fixtures like `temp_dirs`.
+        for attr, assigns in getattr(cls, "setup_assignments", {}).items():
+            val = assigns[-1] if isinstance(assigns, list) and assigns else assigns
+            # find cleanup statements that mention this attr
+            relevant_cleanup = []
+            for stmt in getattr(cls, "teardown_statements", []):
+                try:
+                    if attr in cst.Module(body=[stmt]).code or f"self.{attr}" in cst.Module(body=[stmt]).code:
+                        relevant_cleanup.append(stmt)
+                        if "shutil" in cst.Module(body=[stmt]).code:
+                            needs_shutil = True
+                except Exception:
+                    pass
 
-                def _collect_path_wrapped_attrs(node: Any) -> set[str]:
-                    found: set[str] = set()
-                    if node is None:
-                        return found
-                    # direct Call nodes
-                    if isinstance(node, cst.Call):
-                        fname = _get_callable_name(node.func)
-                        if fname and fname.endswith("Path"):
-                            for a in getattr(node, "args", []) or []:
-                                a_val = getattr(a, "value", None)
-                                if (
-                                    isinstance(a_val, cst.Attribute)
-                                    and isinstance(getattr(a_val, "value", None), cst.Name)
-                                    and getattr(getattr(a_val, "value", None), "value", None) in ("self", "cls")
-                                ):
-                                    aname = getattr(a_val.attr, "value", None)
-                                    if aname:
-                                        found.add(aname)
-                        # recurse into args
-                        for a in getattr(node, "args", []) or []:
-                            av = getattr(a, "value", None)
-                            if isinstance(av, cst.BaseExpression):
-                                found |= _collect_path_wrapped_attrs(av)
-                        return found
-                    # binary ops
-                    if isinstance(node, cst.BinaryOperation):
-                        if isinstance(node.left, cst.BaseExpression):
-                            found |= _collect_path_wrapped_attrs(node.left)
-                        if isinstance(node.right, cst.BaseExpression):
-                            found |= _collect_path_wrapped_attrs(node.right)
-                        return found
-                    # attribute/value containers
-                    if isinstance(node, cst.Attribute) and isinstance(node.value, cst.BaseExpression):
-                        found |= _collect_path_wrapped_attrs(node.value)
-                    if isinstance(node, cst.Subscript) and isinstance(node.value, cst.BaseExpression):
-                        found |= _collect_path_wrapped_attrs(node.value)
-                        for s in node.slice:
-                            inner = getattr(s, "slice", None) or getattr(s, "value", None) or s
-                            if isinstance(inner, cst.BaseExpression):
-                                found |= _collect_path_wrapped_attrs(inner)
-                    if isinstance(node, (cst.List, cst.Tuple, cst.Set, cst.Dict)):
-                        for e in getattr(node, "elements", []) or []:
-                            val = getattr(e, "value", None)
-                            if isinstance(val, cst.BaseExpression):
-                                found |= _collect_path_wrapped_attrs(val)
-                    if isinstance(node, cst.Expr) and isinstance(node.value, cst.BaseExpression):
-                        found |= _collect_path_wrapped_attrs(node.value)
-                    if isinstance(node, cst.Assign) and isinstance(node.value, cst.BaseExpression):
-                        found |= _collect_path_wrapped_attrs(node.value)
+            # Only treat the attribute as yield-style when we find teardown
+            # statements that explicitly reference the attribute (e.g. via
+            # `self.<attr>`). We used to conservatively assume any dir-like
+            # attribute should be yield-style when the class had teardown
+            # statements, but that leads to over-eager binding (creating
+            # local temporaries) for attributes like `config_dir` which are
+            # merely derived from `temp_dir`. Keep the behavior strict:
+            # relevant_cleanup must mention the attribute to force yield-style.
+
+            yield_style = bool(relevant_cleanup)
+            fixture_specs[attr] = SimpleFixtureSpec(
+                name=attr,
+                value_expr=val if isinstance(val, cst.BaseExpression) else None,
+                cleanup_statements=relevant_cleanup,
+                yield_style=yield_style,
+            )
+
+            # generate a trivial fixture node: either yield-style with cleanup or return-style
+            decorator = cst.Decorator(decorator=cst.Attribute(value=cst.Name("pytest"), attr=cst.Name("fixture")))
+
+            # Determine other attribute dependencies referenced as self.<name>
+            def _collect_self_attrs(node: Any) -> set[str]:
+                found: set[str] = set()
+                if node is None or not isinstance(node, cst.BaseExpression):
                     return found
+                if isinstance(node, cst.Attribute):
+                    if isinstance(getattr(node, "value", None), cst.Name) and getattr(node.value, "value", None) in (
+                        "self",
+                        "cls",
+                    ):
+                        if isinstance(node.attr, cst.Name):
+                            found.add(node.attr.value)
+                    # recurse into value and attr
+                    found |= _collect_self_attrs(node.value)
+                    return found
+                if isinstance(node, cst.Call):
+                    found |= _collect_self_attrs(node.func)
+                    for a in getattr(node, "args", []) or []:
+                        inner = getattr(a, "value", None)
+                        found |= _collect_self_attrs(inner)
+                    return found
+                if isinstance(node, cst.Subscript):
+                    found |= _collect_self_attrs(node.value)
+                    for s in getattr(node, "slice", []) or []:
+                        inner = getattr(s, "slice", None) or getattr(s, "value", None) or s
+                        if isinstance(inner, cst.BaseExpression):
+                            found |= _collect_self_attrs(inner)
+                    return found
+                if isinstance(node, (cst.Tuple, cst.List, cst.Set, cst.Dict)):
+                    for e in getattr(node, "elements", []) or []:
+                        val_e = getattr(e, "value", None)
+                        if isinstance(val_e, cst.BaseExpression):
+                            found |= _collect_self_attrs(val_e)
+                    return found
+                # Binary/Comp/Bool
+                if isinstance(node, (cst.BinaryOperation, cst.Comparison, cst.BooleanOperation)):
+                    if hasattr(node, "left") and isinstance(node.left, cst.BaseExpression):
+                        found |= _collect_self_attrs(node.left)
+                    if hasattr(node, "right") and isinstance(node.right, cst.BaseExpression):
+                        found |= _collect_self_attrs(node.right)
+                    for comp in getattr(node, "comparisons", []) or []:
+                        inner = getattr(comp, "comparison", None) or getattr(comp, "operator", None)
+                        if isinstance(inner, cst.BaseExpression):
+                            found |= _collect_self_attrs(inner)
+                    return found
+                # Names and others: nothing
+                return found
 
-                for other_attr, exprs in getattr(cls, "setup_assignments", {}).items():
-                    ve = exprs[-1] if isinstance(exprs, list) and exprs else exprs
-                    try:
-                        path_wrapper_targets |= _collect_path_wrapped_attrs(ve)
-                    except Exception:
-                        # ignore problematic shapes for the prescan
-                        pass
-            except Exception:
-                path_wrapper_targets = set()
+            other_deps = []
+            if isinstance(val, cst.BaseExpression):
+                deps = _collect_self_attrs(val)
+                for name in sorted(deps):
+                    if name != attr:
+                        other_deps.append(name)
 
-            # Build local assignments for each attr using the last recorded
-            # setup expression for that attribute.
-            local_assignments: list[cst.BaseStatement] = []
-            used_names_for_yield: list[tuple[str, str]] = []  # (key, localname)
-            # mapping of attribute -> local variable name for incremental rewrites
-            incremental_mapping: dict[str, str] = {}
-            for attr in attrs:
-                exprs = cls.setup_assignments.get(attr) or []
-                # pick the last assignment expression if multiple
-                value_expr = exprs[-1] if exprs else None
-                # choose a safe local name (avoid collisions)
-                base_local = attr
-                local_name = base_local
-                if local_name in used_local_names:
-                    suffix = 1
-                    while f"{local_name}_{suffix}" in used_local_names:
-                        suffix += 1
-                    local_name = f"{local_name}_{suffix}"
-                used_local_names.add(local_name)
+            # build params for dependencies
+            params: list[cst.Param] = []
+            for d in other_deps:
+                params.append(cst.Param(name=cst.Name(d)))
 
-                # Heuristic: when synthesizing the special "temp_dirs" composite
-                # fixture, only include directory/path-like attributes in the
-                # grouped yield. Non-dir attributes (e.g. main_config,
-                # config_file, mock_* objects) should be emitted as their own
-                # fixtures later so the converter matches the canonical golden
-                # layout.
-                is_dir_like_attr = any(k in attr for k in ("dir", "path", "temp"))
-
-                if fixture_name == "temp_dirs" and not is_dir_like_attr:
-                    # don't create local assignment or include in the grouped
-                    # yield mapping; leave this attr to be emitted as a
-                    # per-attribute fixture later in the loop below.
-                    continue
-
-                # If the expression references previously-created attrs like
-                # temp_dir, rewrite those self.<attr> references to their
-                # local variable names so later local assignments are correct
-                # (e.g., config_dir = Path(temp_dir / "config")).
-                if isinstance(value_expr, cst.BaseExpression):
-                    # Preserve the original value expression here (e.g., a
-                    # tempfile.mkdtemp() call). We only rewrite occurrences of
-                    # self.<attr> to the chosen local name; do not coerce the
-                    # value into a Path(...) wrapper or change tempfile usage.
-                    # transformer to replace self.attr with local variable name
-                    class _ReplaceSelfWithLocal(cst.CSTTransformer):
-                        def __init__(self, mapping: dict[str, str]) -> None:
-                            self.mapping = mapping
-
-                        def leave_Attribute(
-                            self, original: cst.Attribute, updated: cst.Attribute
-                        ) -> cst.BaseExpression:
-                            if isinstance(original.value, cst.Name) and original.value.value in ("self", "cls"):
-                                if isinstance(original.attr, cst.Name):
-                                    attrn = original.attr.value
-                                    if attrn in self.mapping:
-                                        return cst.Name(self.mapping[attrn])
-                            return updated
-
-                    rewritten_value = value_expr.visit(_ReplaceSelfWithLocal(incremental_mapping))
-                    # If this attribute is later referenced inside a Path(...)
-                    # wrapper elsewhere (detected above), prefer to make the
-                    # local binding itself a Path(...) rather than wrapping
-                    # each use. This yields `temp_dir = Path(tempfile.mkdtemp())`
-                    # and then `config_dir = temp_dir / "config"`, matching
-                    # the canonical golden output.
-                    try:
-                        if attr in path_wrapper_targets:
-                            # avoid double-wrapping when already a Path(...) call
-                            fname = None
-                            if isinstance(rewritten_value, cst.Call):
-                                fname = _get_callable_name(rewritten_value.func)
-                            if not (isinstance(rewritten_value, cst.Call) and fname and fname.endswith("Path")):
-                                rewritten_value = cst.Call(
-                                    func=cst.Name("Path"),
-                                    args=[cst.Arg(value=cast(cst.BaseExpression, rewritten_value))],
-                                )
-                    except Exception:
-                        pass
-
-                    # Collapse patterns like Path(temp_dir) / "config" into
-                    # temp_dir / "config" when we've wrapped temp_dir as
-                    # a Path already to avoid double-wrapping.
-                    class _CollapsePathCall(cst.CSTTransformer):
-                        def leave_BinaryOperation(
-                            self, original: cst.BinaryOperation, updated: cst.BinaryOperation
-                        ) -> cst.BaseExpression:
-                            left = updated.left
-                            # check left is Call(Path(...,)) and inner arg is Name
-                            if isinstance(left, cst.Call):
-                                fname = _get_callable_name(left.func)
-                                if fname and fname.endswith("Path") and left.args:
-                                    first_arg = getattr(left.args[0], "value", None)
-                                    if isinstance(first_arg, cst.Name):
-                                        # replace the Call with the bare Name
-                                        new_left = cst.Name(first_arg.value)
-                                        return updated.with_changes(left=new_left)
-                            return updated
-
-                    try:
-                        if isinstance(rewritten_value, cst.BaseExpression):
-                            rewritten_value = cast(cst.BaseExpression, rewritten_value.visit(_CollapsePathCall()))
-                    except Exception:
-                        pass
-                    assign_stmt = cst.SimpleStatementLine(
-                        body=[
-                            cst.Assign(
-                                targets=[cst.AssignTarget(target=cst.Name(local_name))],
-                                value=cast(cst.BaseExpression, rewritten_value),
-                            )
-                        ]
-                    )
-                    local_assignments.append(assign_stmt)
-                # remember mapping for dict yield
-                used_names_for_yield.append((attr, local_name))
-                # record mapping so subsequent attrs can be rewritten
-                incremental_mapping[attr] = local_name
-
-            # create yield dict literal mapping string keys to local names
-            dict_elements: list[cst.DictElement] = []
-            for key, lname in used_names_for_yield:
-                dict_elements.append(cst.DictElement(key=cst.SimpleString(repr(key)), value=cst.Name(lname)))
-            yield_expr = cst.Dict(elements=dict_elements)
-
-            # rewrite cleanup statements to refer to local names instead of self.attr
-            class _AttrRewriterMapping(cst.CSTTransformer):
-                def __init__(self, mapping: dict[str, str]) -> None:
-                    self.mapping = mapping
-
+            # transformer to replace self.X with bare X (dependency param) or local name
+            class _ReplaceSelfWithName(cst.CSTTransformer):
                 def leave_Attribute(self, original: cst.Attribute, updated: cst.Attribute) -> cst.BaseExpression:
                     if isinstance(original.value, cst.Name) and original.value.value in ("self", "cls"):
                         if isinstance(original.attr, cst.Name):
-                            attrn = original.attr.value
-                            if attrn in self.mapping:
-                                return cst.Name(self.mapping[attrn])
+                            name = original.attr.value
+                            return cst.Name(name)
                     return updated
 
-            mapping = {a: local for a, local in used_names_for_yield}
-            safe_cleanup: list[cst.BaseStatement] = []
-            # Only include cleanup statements that reference one of the
-            # attributes included in this composite fixture; otherwise the
-            # cleanup belongs to a different per-attribute fixture and must
-            # not be attached here (avoids moving original_env restores into
-            # the temp_dirs fixture).
-            for stmt in cls.teardown_statements:
-                try:
-                    # Render the statement and include it if it mentions any
-                    # of the attributes in our mapping (e.g., 'self.temp_dir' or
-                    # 'temp_dir'). This mirrors the tolerant fallback used
-                    # elsewhere and avoids false negatives when checking
-                    # statement containers.
-                    rendered = ""
-                    try:
-                        rendered = cst.Module(body=[stmt]).code
-                    except Exception:
-                        rendered = ""
-                    if any((f"self.{a}" in rendered) or (f"{a}" in rendered) for a in mapping):
-                        new_stmt = cast(Any, stmt).visit(_AttrRewriterMapping(mapping))
-                        safe_cleanup.append(new_stmt)
-                        if "shutil" in rendered:
-                            used_shutil = True
-                except Exception:
-                    # conservative: include original if rewrite check fails
-                    safe_cleanup.append(stmt)
-                    try:
-                        rendered = cst.Module(body=[stmt]).code
-                        if "shutil" in rendered:
-                            used_shutil = True
-                    except Exception:
-                        pass
-
-            # build try/yield/finally block
-            try_yield_block = cst.IndentedBlock(body=[cst.SimpleStatementLine(body=[cst.Expr(cst.Yield(yield_expr))])])
-            finalblock_cst = cst.IndentedBlock(body=safe_cleanup or [])
-            # Wrap the finally block in a cst.Finally node (libcst expects a Finally)
-            try_finally = cst.Try(
-                body=try_yield_block, handlers=[], orelse=None, finalbody=cst.Finally(body=finalblock_cst)
-            )
-
-            # assemble function body: locals -> optional mkdirs -> try/finally
-            body_stmts: list[cst.BaseStatement] = []
-            if fixture_name == "temp_dirs":
-                # docstring describing the fixture
-                doc_stmt = cst.SimpleStatementLine(
-                    body=[cst.Expr(cst.SimpleString('"""Create temporary directory structure."""'))]
-                )
-                # include local assignments followed by mkdir calls for dir entries
-                body_stmts.append(doc_stmt)
-                body_stmts.extend(local_assignments)
-                for key, lname in used_names_for_yield:
-                    if key == "temp_dir":
-                        continue
-                    # e.g., config_dir.mkdir(parents=True)
-                    mkdir_call = cst.SimpleStatementLine(
-                        body=[
-                            cst.Expr(
-                                cst.Call(
-                                    func=cst.Attribute(value=cst.Name(lname), attr=cst.Name("mkdir")),
-                                    args=[cst.Arg(keyword=cst.Name("parents"), value=cst.Name("True"))],
+            # collect mkdir/setup statements from recorded setup_methods for this attr
+            setup_stmts: list[cst.BaseStatement] = []
+            for m in getattr(cls, "setup_methods", []) or []:
+                if isinstance(getattr(m, "body", None), cst.IndentedBlock):
+                    for s in getattr(m, "body").body:
+                        try:
+                            code = cst.Module(body=[s]).code
+                        except Exception:
+                            code = ""
+                        # if the parsed statement mentions mkdir for this attr,
+                        # try to transform and preserve it; otherwise synthesize
+                        # a canonical mkdir call only if we detected the pattern
+                        if f"{attr}.mkdir(" in code or f"self.{attr}.mkdir(" in code:
+                            try:
+                                node = s.visit(_ReplaceSelfWithName())
+                                setup_stmts.append(cast(cst.BaseStatement, node))
+                                continue
+                            except Exception:
+                                # fallback to a synthesized mkdir(parents=True)
+                                mkdir_call = cst.SimpleStatementLine(
+                                    body=[
+                                        cst.Expr(
+                                            cst.Call(
+                                                func=cst.Attribute(
+                                                    value=cst.Name(attr if not attr.startswith("self.") else attr),
+                                                    attr=cst.Name("mkdir"),
+                                                ),
+                                                args=[cst.Arg(keyword=cst.Name("parents"), value=cst.Name("True"))],
+                                            )
+                                        )
+                                    ]
                                 )
+                                setup_stmts.append(cast(cst.BaseStatement, mkdir_call.visit(_ReplaceSelfWithName())))
+
+            # decide whether to bind to a local. For literal scalars, prefer
+            # yielding the literal directly so teardown that assigns back to
+            # the attribute remains simple (e.g. "value = None"). For non-
+            # literal expressions, bind to a generated local var and rewrite
+            # cleanup to reference that local.
+            def _is_literal_scalar(node: Any) -> bool:
+                return isinstance(node, (cst.SimpleString, cst.Integer, cst.Float)) or (
+                    isinstance(node, cst.Name) and getattr(node, "value", None) in ("True", "False")
+                )
+
+            # attempt to infer filename literals from local assignments when autocreate
+            try:
+                autocreate = bool(context.get("autocreate", True))
+            except Exception:
+                autocreate = True
+            try:
+                if (
+                    autocreate
+                    and isinstance(val, cst.Call)
+                    and isinstance(val.func, cst.Name)
+                    and val.func.value in ("str",)
+                ):
+                    if val.args:
+                        first = getattr(val.args[0], "value", None)
+                        if isinstance(first, cst.Name):
+                            local_map = getattr(cls, "local_assignments", {}) or {}
+                            mapped = local_map.get(first.value)
+                            if isinstance(mapped, tuple):
+                                assigned_call, idx = mapped
+                            else:
+                                assigned_call, idx = mapped, None
+                            if isinstance(assigned_call, cst.Call):
+                                try:
+                                    if idx is not None and 0 <= idx < len(assigned_call.args):
+                                        cand = getattr(assigned_call.args[idx], "value", None)
+                                        if isinstance(cand, cst.SimpleString):
+                                            val = cand
+                                except Exception:
+                                    pass
+                                for arg_item in getattr(assigned_call, "args", []) or []:
+                                    if (
+                                        getattr(arg_item, "keyword", None)
+                                        and isinstance(arg_item.keyword, cst.Name)
+                                        and arg_item.keyword.value == "filename"
+                                    ):
+                                        cand = getattr(arg_item, "value", None)
+                                        if isinstance(cand, cst.SimpleString):
+                                            val = cand
+                                fname = _get_callable_name(assigned_call.func)
+                                if fname and fname.endswith("Path"):
+                                    for arg_item in getattr(assigned_call, "args", []) or []:
+                                        cand = getattr(arg_item, "value", None)
+                                        if isinstance(cand, cst.SimpleString):
+                                            val = cand
+            except Exception:
+                pass
+
+            if yield_style and isinstance(val, cst.BaseExpression):
+                # attempt to infer filename literals from local assignments when autocreate
+                try:
+                    autocreate = bool(context.get("autocreate", True))
+                except Exception:
+                    autocreate = True
+                # if val is a wrapper like str(sql_file) and autocreate is enabled,
+                # inspect local_assignments to replace the wrapper with the original
+                # literal (filename) when possible.
+                try:
+                    if (
+                        autocreate
+                        and isinstance(val, cst.Call)
+                        and isinstance(val.func, cst.Name)
+                        and val.func.value in ("str",)
+                    ):
+                        # expect first arg to be a Name referencing local
+                        if val.args:
+                            first = getattr(val.args[0], "value", None)
+                            if isinstance(first, cst.Name):
+                                local_map = getattr(cls, "local_assignments", {}) or {}
+                                mapped = local_map.get(first.value)
+                                if isinstance(mapped, tuple):
+                                    assigned_call, idx = mapped
+                                else:
+                                    assigned_call, idx = mapped, None
+                                if isinstance(assigned_call, cst.Call):
+                                    # check positional index
+                                    try:
+                                        if idx is not None and 0 <= idx < len(assigned_call.args):
+                                            cand = getattr(assigned_call.args[idx], "value", None)
+                                            if isinstance(cand, cst.SimpleString):
+                                                val = cand
+                                    except Exception:
+                                        pass
+                                    # check keyword 'filename'
+                                    for arg_item in getattr(assigned_call, "args", []) or []:
+                                        if (
+                                            getattr(arg_item, "keyword", None)
+                                            and isinstance(arg_item.keyword, cst.Name)
+                                            and arg_item.keyword.value == "filename"
+                                        ):
+                                            cand = getattr(arg_item, "value", None)
+                                            if isinstance(cand, cst.SimpleString):
+                                                val = cand
+                                    # path constructor
+                                    fname = _get_callable_name(assigned_call.func)
+                                    if fname and fname.endswith("Path"):
+                                        for arg_item in getattr(assigned_call, "args", []) or []:
+                                            cand = getattr(arg_item, "value", None)
+                                            if isinstance(cand, cst.SimpleString):
+                                                val = cand
+                except Exception:
+                    pass
+
+                # If cleanup references self.<attr> in anything other than a simple
+                # assignment (self.attr = ...), prefer binding. If it's only a
+                # simple assignment we can yield literal directly.
+                # If the attribute was assigned multiple times during setUp,
+                # prefer binding to a local to avoid collisions and make
+                # cleanup rewriting unambiguous (e.g. self.x assigned more
+                # than once). This ensures unique local names are created
+                # for cases like multiple assignments in setUp.
+                must_bind = False
+                try:
+                    if isinstance(assigns, list) and len(assigns) > 1:
+                        must_bind = True
+                except Exception:
+                    pass
+                for s in relevant_cleanup:
+                    try:
+                        code = cst.Module(body=[s]).code.strip()
+                        if f"self.{attr}" in code:
+                            # if it starts with 'self.<attr> =' it's a simple assignment
+                            if not code.startswith(f"self.{attr} ="):
+                                must_bind = True
+                                break
+                    except Exception:
+                        must_bind = True
+                        break
+                # compute conventional local base name and avoid literal-yield
+                # when module-level underscore-prefixed names exist; this
+                # reduces risk of accidental name collisions for generated
+                # locals (e.g., _x_value).
+                base_local = f"_{attr}_value"
+                try:
+                    if any(n.startswith("_") for n in existing_top_names):
+                        must_bind = True
+                except Exception:
+                    pass
+                if _is_literal_scalar(val) and not must_bind:
+                    # yield literal directly, preserve original cleanup statements
+                    yield_stmt = cst.SimpleStatementLine(body=[cst.Expr(cst.Yield(val))])
+                    rewritten_cleanup: list[cst.BaseStatement] = list(relevant_cleanup)
+                    body = cst.IndentedBlock(body=[yield_stmt] + rewritten_cleanup)
+                    func = cst.FunctionDef(
+                        name=cst.Name(attr), params=cst.Parameters(params=params), body=body, decorators=[decorator]
+                    )
+                else:
+                    # Bind yield-style fixtures to a generated local variable so
+                    # tearDown references can be rewritten safely. This avoids
+                    # yielding literals directly which complicates cleanup.
+                    # ensure unique local var name
+                    base_local = f"_{attr}_value"
+                    local_var = base_local
+                    i = 1
+                    # avoid collisions with module-level/top-level names
+                    while local_var in locals() or local_var in globals() or local_var in existing_top_names:
+                        local_var = f"{base_local}_{i}"
+                        i += 1
+
+                    # create assignment to local_var
+                    assign = cst.SimpleStatementLine(
+                        body=[
+                            cst.Assign(
+                                targets=[cst.AssignTarget(target=cst.Name(local_var))],
+                                value=cast(cst.BaseExpression, val.visit(_ReplaceSelfWithName())),
                             )
                         ]
                     )
-                    body_stmts.append(mkdir_call)
-            else:
-                body_stmts.extend(local_assignments)
-            body_stmts.append(try_finally)
-            body = cst.IndentedBlock(body=body_stmts)
 
-            # ensure typing imports for Generator, Dict and Path
-            all_typing_needed.update({"Generator", "Dict", "Path", "Any"})
+                    # yield the local_var
+                    yield_stmt = cst.SimpleStatementLine(body=[cst.Expr(cst.Yield(cst.Name(local_var)))])
 
-            # Build a precise return annotation: Generator[Dict[str, Path], None, None]
-            # Annotation AST: Generator[Dict[str, Path], None, None]
-            # libcst representation: Subscript(Name('Generator'), [SubscriptElement(Index(Subscript(Name('Dict'), [SubscriptElement(Index(Name('str'))), SubscriptElement(Index(Name('Path')))]))), SubscriptElement(Index(Name('None'))), SubscriptElement(Index(Name('None')))])
-            dict_inner = cst.Subscript(
-                value=cst.Name("Dict"),
-                slice=[
-                    cst.SubscriptElement(slice=cst.Index(value=cst.Name("str"))),
-                    cst.SubscriptElement(slice=cst.Index(value=cst.Name("Path"))),
-                ],
-            )
-            gen_sub = cst.Subscript(
-                value=cst.Name("Generator"),
-                slice=[
-                    cst.SubscriptElement(slice=cst.Index(value=dict_inner)),
-                    cst.SubscriptElement(slice=cst.Index(value=cst.Name("None"))),
-                    cst.SubscriptElement(slice=cst.Index(value=cst.Name("None"))),
-                ],
-            )
+                    # rewrite cleanup to reference the local_var instead of self.attr
+                    rewritten_cleanup = []
+                    for s in relevant_cleanup:
+                        new_s = cast(cst.BaseStatement, s.visit(_ReplaceSelfWithName()))
 
-            return_ann = cst.Annotation(annotation=gen_sub)
+                        class _ReplaceAttrWithLocal(cst.CSTTransformer):
+                            def leave_Name(self, original: cst.Name, updated: cst.Name) -> cst.BaseExpression:
+                                if original.value == attr:
+                                    return cst.Name(local_var)
+                                return updated
 
-            # create decorator and function def with precise return annotation
-            decorator = cst.Decorator(decorator=cst.Attribute(value=cst.Name("pytest"), attr=cst.Name("fixture")))
-            func = cst.FunctionDef(
-                name=cst.Name(fixture_name),
-                params=cst.Parameters(),
-                body=body,
-                decorators=[decorator],
-                returns=return_ann,
-            )
-            fixture_nodes.append(func)
-            # record a simple spec entry for completeness
-            specs[fixture_name] = FixtureSpec(
-                name=fixture_name, value_expr=None, cleanup_statements=safe_cleanup, yield_style=True
-            )
-            # mark attributes included in this composite fixture as handled
-            for a, _ in used_names_for_yield:
-                handled_attrs.add(a)
-        # Detect composite helper-call patterns recorded in collector.local_assignments.
-        # If multiple local names map to the same Call (tuple-unpack), and
-        # corresponding attributes are assigned from those locals, synthesize
-        # a single composite fixture that calls the helper and returns the
-        # tuple, wiring upstream fixtures for any self.<attr> arguments.
-        # reuse the handled_attrs from above (do not reinitialize)
-        try:
-            # Build mapping from call source code -> list[(local_name, index, call_node)]
-            call_groups: dict[str, list[tuple[str, Optional[int], Any]]] = {}
-            for local_name, val in getattr(cls, "local_assignments", {}).items():
-                # val is (assigned_value, maybe_index)
-                assigned_call = None
-                idx = None
+                        new_s = cast(cst.BaseStatement, new_s.visit(_ReplaceAttrWithLocal()))
+                        rewritten_cleanup.append(new_s)
+
+                    include_setup = not bool(getattr(cls, "setup_methods", []))
+                    body = cst.IndentedBlock(
+                        body=([assign] if include_setup else [assign]) + [yield_stmt] + rewritten_cleanup
+                    )
+                    func = cst.FunctionDef(
+                        name=cst.Name(attr), params=cst.Parameters(params=params), body=body, decorators=[decorator]
+                    )
+            elif not yield_style and isinstance(val, cst.BaseExpression):
+                # rewrite self.* references in the return expression to params
+                val_simple = cast(cst.BaseExpression, val.visit(_ReplaceSelfWithName()))
+                return_stmt = cst.SimpleStatementLine(body=[cst.Return(val_simple)])
+
+                # ensure mkdir/setup statements are preserved even for return-style
+                # fixtures: detect mkdir calls in recorded setup_methods and
+                # append transformed or synthesized mkdir statements.
+                preserved_setup: list[cst.BaseStatement] = list(setup_stmts)
+                for m in getattr(cls, "setup_methods", []) or []:
+                    if isinstance(getattr(m, "body", None), cst.IndentedBlock):
+                        for s in getattr(m, "body").body:
+                            try:
+                                code = cst.Module(body=[s]).code
+                            except Exception:
+                                code = ""
+                            if f"{attr}.mkdir(" in code or f"self.{attr}.mkdir(" in code:
+                                try:
+                                    node = s.visit(_ReplaceSelfWithName())
+                                    preserved_setup.append(cast(cst.BaseStatement, node))
+                                except Exception:
+                                    mkdir_call = cst.SimpleStatementLine(
+                                        body=[
+                                            cst.Expr(
+                                                cst.Call(
+                                                    func=cst.Attribute(
+                                                        value=cst.Name(attr if not attr.startswith("self.") else attr),
+                                                        attr=cst.Name("mkdir"),
+                                                    ),
+                                                    args=[cst.Arg(keyword=cst.Name("parents"), value=cst.Name("True"))],
+                                                )
+                                            )
+                                        ]
+                                    )
+                                    preserved_setup.append(
+                                        cast(cst.BaseStatement, mkdir_call.visit(_ReplaceSelfWithName()))
+                                    )
+
+                include_setup = not bool(getattr(cls, "setup_methods", []))
+                body = cst.IndentedBlock(body=(preserved_setup if include_setup else []) + [return_stmt])
+                # try to infer a precise return annotation for this fixture
                 try:
-                    assigned_call, idx = val
+                    ann_res, names_req = _infer_ann(val)
+                    needs_typing.update(names_req)
+                    # Only emit a return annotation when it's more specific than Any.
+                    ann_node = getattr(ann_res, "annotation", None)
+                    # Emit return annotations when they are useful and stable.
+                    # Allow simple Name annotations (e.g., "int", "Path") and
+                    # Tuple[...] subscript annotations (commonly inferred for
+                    # tuple literals). Skip other complex/subscript annotations
+                    # to avoid noisy goldens.
+                    if isinstance(ann_node, cst.Name):
+                        if getattr(ann_node, "value", None) == "Any":
+                            func = cst.FunctionDef(
+                                name=cst.Name(attr),
+                                params=cst.Parameters(params=params),
+                                body=body,
+                                decorators=[decorator],
+                            )
+                        else:
+                            func = cst.FunctionDef(
+                                name=cst.Name(attr),
+                                params=cst.Parameters(params=params),
+                                body=body,
+                                decorators=[decorator],
+                                returns=ann_res,
+                            )
+                    elif (
+                        isinstance(ann_node, cst.Subscript)
+                        and isinstance(getattr(ann_node, "value", None), cst.Name)
+                        and getattr(ann_node.value, "value", None) in ("Tuple", "List")
+                    ):
+                        # allow Tuple[...] annotations for tuple literals
+                        func = cst.FunctionDef(
+                            name=cst.Name(attr),
+                            params=cst.Parameters(params=params),
+                            body=body,
+                            decorators=[decorator],
+                            returns=ann_res,
+                        )
+                    else:
+                        func = cst.FunctionDef(
+                            name=cst.Name(attr), params=cst.Parameters(params=params), body=body, decorators=[decorator]
+                        )
                 except Exception:
-                    assigned_call = val
-                    idx = None
-                if isinstance(assigned_call, cst.Call):
-                    try:
-                        # rendering a Call requires wrapping it into an Expr
-                        key = cst.Module(body=[cst.SimpleStatementLine(body=[cst.Expr(assigned_call)])]).code
-                    except Exception:
-                        # fallback to repr-based key if rendering fails
-                        key = repr(assigned_call)
-                    call_groups.setdefault(key, []).append((local_name, idx, assigned_call))
+                    func = cst.FunctionDef(
+                        name=cst.Name(attr), params=cst.Parameters(params=params), body=body, decorators=[decorator]
+                    )
+            else:
+                # fallback: return None
+                return_stmt = cst.SimpleStatementLine(body=[cst.Return(cst.Name("None"))])
+                body = cst.IndentedBlock(body=[return_stmt])
+                func = cst.FunctionDef(name=cst.Name(attr), params=cst.Parameters(), body=body, decorators=[decorator])
 
-            # For each group with multiple locals, check whether those locals
-            # are later mapped into class attributes via setup_assignments
+            fixture_nodes.append(func)
+
+            # add typing names inferred from value expression
+            try:
+                if isinstance(val, cst.BaseExpression):
+                    ann_res, names_req = _infer_ann(val)
+                    needs_typing.update(names_req)
+            except Exception:
+                pass
+
+                # minimal typing inference: if dict literal, add typing names.
+                # For sequence/tuple/set literals we rely on _infer_ann above to
+                # provide precise typing requirements. Avoid unconditionally
+                # adding 'Any' here because it pollutes generated imports and
+                # golden outputs when a precise annotation (e.g., Tuple[str,int])
+                # was inferred.
+            try:
+                if isinstance(val, (cst.Dict,)):
+                    needs_typing.update({"Dict"})
+                # detect Path(...) usage in expressions
+                if isinstance(val, cst.Call):
+                    fname = None
+                    if isinstance(val.func, cst.Name):
+                        fname = val.func.value
+                    elif isinstance(val.func, cst.Attribute):
+                        fname = getattr(val.func.attr, "value", None)
+                    if fname == "Path" or (isinstance(fname, str) and fname.endswith("Path")):
+                        needs_typing.add("Path")
+            except Exception:
+                pass
+
+    # If any fixture was yield-style, we need Generator imported
+    try:
+        if any(s.yield_style for s in fixture_specs.values()):
+            needs_typing.add("Generator")
+    except Exception:
+        pass
+
+    # If setUp contains tuple-unpacking from a helper call (recorded in
+    # local_assignments), synthesize a NamedTuple container and a bundled
+    # fixture that constructs and yields it. This is narrower than the old
+    # composite heuristics: we only bundle when multiple locals come from the
+    # same Call and those locals map to attributes recorded in setup_assignments.
+    try:
+        used_names: set[str] = set()
+        for cls_name, cls in out.classes.items():
+            local_map = getattr(cls, "local_assignments", {}) or {}
+            # group by the assigned call expression source
+            call_groups: dict[str, list[tuple[str, int, Any]]] = {}
+            for local_name, val in local_map.items():
+                assigned_call, idx = val if isinstance(val, tuple) else (val, None)
+                if not isinstance(assigned_call, cst.Call):
+                    continue
+                try:
+                    key = cst.Module(body=[cst.SimpleStatementLine(body=[cst.Expr(assigned_call)])]).code
+                except Exception:
+                    key = repr(assigned_call)
+                call_groups.setdefault(key, []).append((local_name, idx, assigned_call))
+
             for group in call_groups.values():
                 if len(group) < 2:
                     continue
-                # find which locals correspond to attributes
+                # map locals to attributes if possible
                 local_to_attr: dict[str, str] = {}
                 for local_name, idx, assigned_call in group:
-                    # search for attributes whose value_expr references this local
-                    for attr_name, assigns in cls.setup_assignments.items():
-                        # take the last assignment as effective
+                    for attr_name, assigns in getattr(cls, "setup_assignments", {}).items():
                         v = assigns[-1] if isinstance(assigns, list) and assigns else assigns
-                        # common pattern: self.attr = Name(local_name) or Call(Name(local_name))
-                        # Match patterns where the attribute is assigned from a local name
-                        # Direct name: self.attr = local_name
                         if isinstance(v, cst.Name) and v.value == local_name:
                             local_to_attr[local_name] = attr_name
-                        # If the local variable name equals the attribute name, assume they correspond
-                        # (covers patterns where the author rebinds or assigns literals to same-name attrs)
                         elif local_name == attr_name:
                             local_to_attr[local_name] = attr_name
-                        # Call wrappers: self.attr = wrapper(local_name)
                         elif isinstance(v, cst.Call) and v.args:
                             for arg_item in v.args:
                                 a_val = getattr(arg_item, "value", None)
                                 if isinstance(a_val, cst.Name) and a_val.value == local_name:
                                     local_to_attr[local_name] = attr_name
                                     break
+
                 if not local_to_attr:
                     continue
 
-                # All right: synthesize composite fixture; prefer bundling into
-                # a NamedTuple-based container and a single yield-style fixture
-                # derived from the TestCase class name.
-                assigned_call = group[0][2]
-
-                # derive a container and fixture name from the TestCase class
+                # derive container and fixture names from TestCase class name
                 def _derive_names(class_name: str) -> tuple[str, str]:
                     base = class_name
                     if base.startswith("Test") and len(base) > 4:
                         base = base[4:]
                     if not base:
                         base = class_name
-                    # NamedTuple name (private)
                     named = f"_{base}Data"
-                    # fixture name: snake_case + _data
-                    # Use a two-step regex to handle acronyms (e.g., API -> api)
                     import re
 
                     s1 = re.sub(r"([A-Z]+)([A-Z][a-z])", r"\1_\2", base)
@@ -850,91 +837,31 @@ def generator_stage(context: dict[str, Any]) -> dict[str, Any]:
                     return named, fixture_nm
 
                 namedtuple_name, fixture_name = _derive_names(cls_name)
+                # ensure uniqueness for namedtuple and fixture against module names
+                if namedtuple_name in used_names or namedtuple_name in existing_top_names:
+                    # skip if conflicting
+                    continue
+                used_names.add(namedtuple_name)
+                # avoid fixture name colliding with existing top-level names
+                if fixture_name in existing_top_names:
+                    suffix_i = 1
+                    base = fixture_name
+                    while f"{base}_{suffix_i}" in existing_top_names:
+                        suffix_i += 1
+                    fixture_name = f"{base}_{suffix_i}"
 
-                # ensure fixture_name doesn't collide with module-level names
-                def _unique_name(base: str, taken: set[str]) -> str:
-                    if base not in taken:
-                        taken.add(base)
-                        return base
-                    i = 1
-                    while True:
-                        cand = f"{base}_{i}"
-                        if cand not in taken:
-                            taken.add(cand)
-                            return cand
-                        i += 1
-
-                fixture_name = _unique_name(fixture_name, used_local_names)
-
-                # prepare to emit NamedTuple definition and fixture
-                # We'll collect required typing names into `typing_needed` and
-                # accumulate them into `all_typing_needed` (declared outside the
-                # class loop) so the centralized import injector stage can insert
-                # a single `from typing import ...` line.
-                typing_needed: set[str] = {"NamedTuple", "List", "Dict", "Any"}
-
-                # determine upstream fixture parameters: collect referenced self.<name> attrs
-                attach_params: list[cst.Param] = []
-                upstream_names: list[str] = []
-                for arg in getattr(assigned_call, "args", []) or []:
-                    a_val = getattr(arg, "value", None)
-                    if (
-                        isinstance(a_val, cst.Attribute)
-                        and isinstance(getattr(a_val, "value", None), cst.Name)
-                        and getattr(getattr(a_val, "value", None), "value", None) in ("self", "cls")
-                    ):
-                        aname = getattr(a_val.attr, "value", None)
-                        if aname and aname not in upstream_names:
-                            upstream_names.append(aname)
-                            attach_params.append(cst.Param(name=cst.Name(aname)))
-
-                # prepare call expression with self.X replaced by bare names
-                class _ReplaceSelf(cst.CSTTransformer):
-                    def leave_Attribute(self, original: cst.Attribute, updated: cst.Attribute) -> cst.BaseExpression:
-                        if isinstance(original.value, cst.Name) and original.value.value in ("self", "cls"):
-                            if isinstance(original.attr, cst.Name):
-                                return cst.Name(original.attr.value)
-                        return updated
-
-                call_in_fixture = assigned_call.visit(_ReplaceSelf())
-
-                # build tuple assignment target list in order of indices
-                # sort group entries by index when present
-                sorted_group = sorted(group, key=lambda x: (x[1] if x[1] is not None else 0))
-                targets = [cst.Name(local) for local, _, _ in sorted_group]
-                assign_target = cst.Assign(
-                    targets=[cst.AssignTarget(target=cst.Tuple(elements=[cst.Element(value=t) for t in targets]))],
-                    value=call_in_fixture,
-                )
-
-                # Build NamedTuple fields and fixture body that constructs and yields it
-                # Create class body: docstring + annotated assignments
+                # build a minimal NamedTuple class with Any annotations
                 class_body: list[cst.BaseStatement] = []
-                # docstring
                 class_body.append(
                     cst.SimpleStatementLine(
                         body=[cst.Expr(cst.SimpleString('"""Container for test data and resources."""'))]
                     )
                 )
-                # fields: derive from local_to_attr mapping order
+                sorted_group = sorted(group, key=lambda x: (x[1] if x[1] is not None else 0))
                 for local, _, _ in sorted_group:
-                    attr = local_to_attr.get(local)
-                    # best-effort type: infer from recorded setup assignment if present
-                    ann_node: cst.Annotation
-                    orig_val = None
-                    if attr:
-                        setup_assigns: Optional[list[Any]] = cls.setup_assignments.get(attr)
-                        if setup_assigns:
-                            orig_val = setup_assigns[-1]
-                    if orig_val is not None and isinstance(orig_val, cst.BaseExpression):
-                        ann_node, names_needed = _infer_ann(orig_val)
-                    else:
-                        ann_node, names_needed = _infer_ann(None)
-                    # collect typing names for this NamedTuple
-                    for n in names_needed:
-                        typing_needed.add(n)
+                    attr = local_to_attr.get(local, local)
                     ann_assign = cst.AnnAssign(
-                        target=cst.Name(attr if attr else local), annotation=ann_node, value=None
+                        target=cst.Name(attr), annotation=cst.Annotation(annotation=cst.Name("Any")), value=None
                     )
                     class_body.append(cst.SimpleStatementLine(body=[ann_assign]))
 
@@ -944,627 +871,69 @@ def generator_stage(context: dict[str, Any]) -> dict[str, Any]:
                     body=cst.IndentedBlock(body=class_body),
                 )
 
-                # fixture body: assign call result to locals, then yield container and finally cleanup
-                assign_stmt = cst.SimpleStatementLine(body=[assign_target])
-                # build TestData constructor call args
-                ctor_args: list[cst.Arg] = []
-                for local, _, _ in sorted_group:
-                    attr = local_to_attr.get(local)
-                    # preserve wrappers when present
-                    orig_v: Optional[Any] = None
-                    if attr:
-                        setup_assigns = cls.setup_assignments.get(attr)
-                        if setup_assigns:
-                            orig_v = setup_assigns[-1]
-                    if isinstance(orig_v, cst.Call) and orig_v.args:
-                        new_call = orig_v.with_changes(args=[cst.Arg(value=cst.Name(local))])
-                        ctor_args.append(cast(cst.Arg, cst.Arg(value=new_call)))
-                    else:
-                        ctor_args.append(cast(cst.Arg, cst.Arg(value=cst.Name(local))))
+                # emit fixture that assigns call result to locals and yields the NamedTuple
+                assigned_call = group[0][2]
 
-                # build yield expression: yield NamedTupleName(field=...)
-                ctor = cst.Call(func=cst.Name(namedtuple_name), args=ctor_args)
-                yield_stmt = cst.SimpleStatementLine(body=[cst.Expr(cst.Yield(ctor))])
-
-                # map teardown statements: rewrite self.attr -> local name
-                rewritten_cleanup: list[cst.BaseStatement] = []
-                for stmt in cls.teardown_statements:
-                    new_stmt = cast(cst.BaseStatement, cast(Any, stmt).visit(_ReplaceSelf()))
-                    rewritten_cleanup.append(new_stmt)
-
-                # finally block: put cleanup statements
-                finally_block: list[cst.BaseStatement] = []
-                for s in rewritten_cleanup:
-                    finally_block.append(cast(cst.BaseStatement, s))
-
-                # construct try/finally as code: try: <assign>; yield; finally: <cleanup>
-                if finally_block:
-                    try_block = cst.Try(
-                        body=cst.IndentedBlock(body=[assign_stmt, yield_stmt]),
-                        handlers=[],
-                        orelse=None,
-                        finalbody=cst.Finally(body=cst.IndentedBlock(body=list(finally_block))),
-                    )
-                    body = cst.IndentedBlock(body=[try_block])
-                    func = cst.FunctionDef(
-                        name=cst.Name(fixture_name),
-                        params=cst.Parameters(params=attach_params),
-                        body=body,
-                        decorators=[
-                            cst.Decorator(decorator=cst.Attribute(value=cst.Name("pytest"), attr=cst.Name("fixture")))
-                        ],
-                    )
-                else:
-                    # no cleanup: simpler non-yield fixture that returns the container
-                    return_stmt = cst.SimpleStatementLine(body=[cst.Return(ctor)])
-                    body = cst.IndentedBlock(body=[assign_stmt, return_stmt])
-                    func = cst.FunctionDef(
-                        name=cst.Name(fixture_name),
-                        params=cst.Parameters(params=attach_params),
-                        body=body,
-                        decorators=[
-                            cst.Decorator(decorator=cst.Attribute(value=cst.Name("pytest"), attr=cst.Name("fixture")))
-                        ],
-                    )
-
-                # aggregate typing names for later injection by import_injector
-                all_typing_needed.update(typing_needed)
-                # add class_def and fixture
-                fixture_nodes.append(class_def)
-                fixture_nodes.append(func)
-
-                # mark attributes as handled so we skip generating per-attr fixtures
-                for local_name in local_to_attr:
-                    handled_attrs.add(local_to_attr[local_name])
-        except Exception:
-            # be conservative: on any unexpected shape, ignore and fall back
-            handled_attrs = set()
-        for attr, value in cls.setup_assignments.items():
-            if attr in handled_attrs:
-                # skip attributes already satisfied by a composite fixture
-                continue
-            # collector may record multiple assignments per attribute as a list
-            multi_assigned = False
-            if isinstance(value, list):
-                multi_assigned = len(value) > 1
-                # use the last assigned value as the effective value
-                value_expr = value[-1] if value else None
-            else:
-                value_expr = value
-            fname = f"{attr}"
-            # If the recorded value is a self-referential placeholder (or a
-            # simple wrapper like str(sql_file)) and the test author also
-            # provided a sibling '<prefix>_content' assignment, auto-generate
-            # a tmp_path-based fixture that writes that content to a file and
-            # returns its string path. This avoids emitting broken
-            # self-referential placeholders like `str(sql_file)` in fixtures.
-            try:
-                # conservative check: only for attrs that look like file names
-                # and only if autocreate is enabled via CLI or API.
-                if autocreate_flag and isinstance(attr, str) and attr.endswith("_file"):
-                    prefix = attr[: -len("_file")]
-                    candidate = f"{prefix}_content"
-                    if candidate in cls.setup_assignments:
-                        # attempt to infer filename from local assignments recorded
-                        inferred_filename = None
-                        # Collector may have recorded local assignments mapping names to (expr, index)
-                        local_map = getattr(cls, "local_assignments", {}) or {}
-
-                        def _get_callable_name(node: Any) -> Optional[str]:
-                            """Return a dotted name for simple Name/Attribute callables, or None.
-
-                            Examples: Name('Path') -> 'Path', Attribute(Name('pathlib'), Name('Path')) -> 'pathlib.Path'
-                            """
-                            if node is None:
-                                return None
-                            if isinstance(node, cst.Name):
-                                return node.value
-                            if isinstance(node, cst.Attribute):
-                                parts: list[str] = []
-                                cur: Any = node
-                                # unwind attributes defensively
-                                while isinstance(cur, cst.Attribute):
-                                    if isinstance(getattr(cur, "attr", None), cst.Name):
-                                        parts.append(cur.attr.value)
-                                    val = getattr(cur, "value", None)
-                                    if isinstance(val, cst.Name):
-                                        parts.append(val.value)
-                                        break
-                                    # prepare to unwrap further or bail
-                                    cur = val
-                                return ".".join(reversed(parts)) if parts else None
-                            return None
-
-                        def _string_from_simple(aval: Any) -> Optional[str]:
-                            if isinstance(aval, cst.SimpleString):
-                                return aval.value.strip("\"'")
-                            return None
-
-                        def _find_filename_from_call(call: cst.Call) -> Optional[str]:
-                            # 1) check keyword args commonly used for filenames
-                            for arg in getattr(call, "args", []) or []:
-                                # keyword may be Name (keyword arg) or None (positional)
-                                kw = getattr(arg, "keyword", None)
-                                if (
-                                    kw
-                                    and isinstance(kw, cst.Name)
-                                    and kw.value.lower() in ("filename", "name", "path", "file")
-                                ):
-                                    sval = _string_from_simple(getattr(arg, "value", None))
-                                    if sval:
-                                        return sval
-                            # 2) check positional args for a string literal-like filename
-                            for arg in getattr(call, "args", []) or []:
-                                sval = _string_from_simple(getattr(arg, "value", None))
-                                if sval:
-                                    return sval
-                            # 3) check for Path(...) constructions: if the callee looks like Path
-                            func_name = _get_callable_name(call.func)
-                            if func_name and func_name.endswith("Path"):
-                                for arg in getattr(call, "args", []) or []:
-                                    sval = _string_from_simple(getattr(arg, "value", None))
-                                    if sval:
-                                        return sval
-                            return None
-
-                        # value_expr may be a Call wrapping a Name referencing a local var (e.g., str(sql_file))
-                        target_name = None
-                        if isinstance(value, list):
-                            ve = value[-1] if value else None
-                        else:
-                            ve = value
-                        if isinstance(ve, cst.Call) and ve.args:
-                            for arg_item in ve.args:
-                                a_val = getattr(arg_item, "value", None)
-                                if isinstance(a_val, cst.Name):
-                                    target_name = a_val.value
-                                    break
-
-                        # if we found a local assignment for that name, inspect its call args
-                        if target_name and target_name in local_map:
-                            assigned_call, index = local_map[target_name]
-                            if isinstance(assigned_call, cst.Call):
-                                inferred_filename = _find_filename_from_call(assigned_call)
-
-                        # fallback filename if none inferred
-                        if not inferred_filename:
-                            inferred_filename = f"{attr}.sql"
-
-                        # create autocreated fixture and skip normal fixture generation
-                        fixture_node = create_autocreated_file_fixture(attr, candidate, filename=inferred_filename)
-                        specs[fname] = FixtureSpec(
-                            name=fname,
-                            value_expr=value if not isinstance(value, list) else (value[-1] if value else None),
-                            cleanup_statements=[],
-                            yield_style=False,
-                        )
-                        fixture_nodes.append(fixture_node)
-                        continue
-            except Exception:
-                # be conservative: fall back to normal behavior on unexpected shapes
-                pass
-            # find teardown statements that reference this attr
-            # We accept a mix of libcst statement flavors here; widen to Any to
-            # avoid variance and typeshed mismatches while preserving runtime
-            # behavior. We'll narrow later at stage boundaries where needed.
-            relevant_cleanup: list[Any] = []
-            for stmt in cls.teardown_statements:
-                # inspect common containers similarly to legacy
-                def _stmt_references(s: Any) -> bool:
-                    # SimpleStatementLine: inspect expr or assign
-                    if isinstance(s, cst.SimpleStatementLine) and s.body:
-                        expr = s.body[0]
-                        # assignment
-                        if isinstance(expr, cst.Assign):
-                            for t in expr.targets:
-                                target = getattr(t, "target", t)
-                                if _references_attribute(target, attr):
-                                    return True
-                            if _references_attribute(expr.value, attr):
-                                return True
-                        # expression wrapper (e.g., Expr(Call(...)))
-                        if isinstance(expr, cst.Expr):
-                            if _references_attribute(expr.value, attr):
-                                return True
-                        # delete statements: del self.attr
-                        # Some libcst versions/typeshed don't expose a Delete symbol
-                        # that mypy recognizes. Detect by class name to avoid mypy errors.
-                        cls = getattr(expr, "__class__", None)
-                        if cls is not None and getattr(cls, "__name__", None) == "Delete":
-                            for t in getattr(expr, "targets", []):
-                                target = getattr(t, "target", t)
-                                if _references_attribute(target, attr):
-                                    return True
-                    # If/IndentedBlock: inspect body and orelse
-                    if isinstance(s, cst.If):
-                        if _references_attribute(s.test, attr):
-                            return True
-                        for inner in getattr(s.body, "body", []):
-                            if _stmt_references(inner):
-                                return True
-                        orelse = getattr(s, "orelse", None)
-                        if orelse:
-                            if isinstance(orelse, cst.IndentedBlock):
-                                for inner in getattr(orelse, "body", []):
-                                    if _stmt_references(inner):
-                                        return True
-                            elif isinstance(orelse, cst.If):
-                                if _stmt_references(orelse):
-                                    return True
-                    # IndentedBlock
-                    if isinstance(s, cst.IndentedBlock):
-                        for inner in getattr(s, "body", []):
-                            if _stmt_references(inner):
-                                return True
-                    return False
-
-                try:
-                    if _stmt_references(stmt):
-                        # cast to the wider union to satisfy list element typing
-                        relevant_cleanup.append(stmt)
-                except Exception:
-                    # be conservative: ignore unexpected shapes
-                    pass
-
-            # fallback: if our structural checks missed some unusual Delete/cleanup
-            # forms, conservatively include statements whose rendered code contains
-            # both 'del' and the attribute name. This mirrors the legacy transformer's
-            # tolerant behavior and avoids missing cleanup like 'del self.x'.
-            if not relevant_cleanup:
-                for stmt in cls.teardown_statements:
-                    try:
-                        rendered = cst.Module(body=[stmt]).code
-                        if "del" in rendered and (f"self.{attr}" in rendered or f"{attr}" in rendered):
-                            relevant_cleanup.append(stmt)
-                    except Exception:
-                        # ignore rendering issues; keep conservative behavior
-                        pass
-
-            has_cleanup = bool(relevant_cleanup)
-            yield_style = has_cleanup
-            spec = FixtureSpec(
-                name=fname, value_expr=value_expr, cleanup_statements=relevant_cleanup.copy(), yield_style=yield_style
-            )
-            specs[fname] = spec
-            # infer typing requirements from the recorded value expression so
-            # import_injector can insert needed typing imports even for
-            # non-composite fixtures (e.g., list/dict/set literals)
-            try:
-                ann_infer, names_req = (
-                    _infer_ann(value_expr) if isinstance(value_expr, cst.BaseExpression) else _infer_ann(None)
-                )
-                for n in names_req:
-                    all_typing_needed.add(n)
-            except Exception:
-                # be conservative: ignore inference failures
-                pass
-            # create a minimal fixture node: @pytest.fixture
-            decorator = cst.Decorator(decorator=cst.Attribute(value=cst.Name("pytest"), attr=cst.Name("fixture")))
-
-            # determine local name and ensure uniqueness
-            def _choose_local_name(base: str, taken: set[str]) -> str:
-                """Deterministically pick a unique local name by appending
-                a numeric suffix when needed. Returns the chosen name and
-                reserves it in `taken`.
-                """
-                if base not in taken:
-                    taken.add(base)
-                    return base
-                suffix = 1
-                while True:
-                    candidate = f"{base}_{suffix}"
-                    if candidate not in taken:
-                        taken.add(candidate)
-                        return candidate
-                    suffix += 1
-
-            base_local = f"_{attr}_value"
-            local_name = _choose_local_name(base_local, used_local_names)
-            spec.local_value_name = local_name
-
-            # decide whether to bind to a local variable or return the value
-            # directly. For non-yield fixtures where the value is a simple
-            # literal, container literal, or comprehension, return the value
-            # directly to produce a canonical simple fixture form. Otherwise,
-            # bind to a unique local name so cleanup rewrites can reference it.
-            def _is_simple_value(expr: Any) -> bool:
-                if expr is None:
-                    return False
-                # simple scalar literals
-                if isinstance(expr, (cst.Integer, cst.Float, cst.Imaginary, cst.SimpleString, cst.Name)):
-                    return True
-                # container literals
-                if isinstance(expr, (cst.List, cst.Tuple, cst.Set, cst.Dict)):
-                    return True
-                # comprehensions
-                cls = getattr(expr, "__class__", None)
-                if cls is not None and getattr(cls, "__name__", "").endswith("Comp"):
-                    return True
-                return False
-
-            # previously we used local temporaries here; they were removed
-            # to avoid unused-variable lints. The direct-return decision is
-            # performed in-place where needed.
-
-            # helper rewriter to replace self.attr or cls.attr with local_name
-            class _AttrRewriterTarget(cst.CSTTransformer):
-                def __init__(self, target_attr: str, local: str) -> None:
-                    self.target_attr = target_attr
-                    self.local = local
-
-                def leave_Attribute(self, original: cst.Attribute, updated: cst.Attribute) -> cst.BaseExpression:
-                    if isinstance(original.value, cst.Name) and original.value.value in ("self", "cls"):
-                        if isinstance(original.attr, cst.Name) and original.attr.value == self.target_attr:
-                            return cst.Name(self.local)
-                    return updated
-
-            # body: return or yield
-            # skip creating fixture if a top-level function with same name already exists
-            if module is not None and any(
-                isinstance(n, cst.FunctionDef) and n.name.value == fname for n in module.body
-            ):
-                # still record spec but don't create a duplicate fixture node
-                specs[fname] = spec
-                continue
-
-            if yield_style:
-                # If the value is a simple literal and all cleanup statements are
-                # simple assignments or deletions targeting the attribute, we
-                # can yield the literal directly and rewrite cleanup to use the
-                # fixture name (e.g., value = None). Otherwise, bind to a
-                # unique local name and rewrite cleanup to reference that
-                # local name so complex control flow works safely.
-                def _is_simple_cleanup_statement(s: Any) -> bool:
-                    # Simple assignment like `self.attr = X` or `del self.attr`
-                    if isinstance(s, cst.SimpleStatementLine) and s.body:
-                        expr = s.body[0]
-                        if isinstance(expr, cst.Assign):
-                            target = expr.targets[0].target
-                            if (
-                                isinstance(target, cst.Attribute)
-                                and isinstance(target.value, cst.Name)
-                                and target.value.value in ("self", "cls")
-                            ):
-                                return True
-                        # Some libcst versions/typeshed don't expose a Delete symbol
-                        # that mypy recognizes. Detect Delete by class name to avoid
-                        # mypy attr-defined errors.
-                        cls = getattr(expr, "__class__", None)
-                        if cls is not None and getattr(cls, "__name__", None) == "Delete":
-                            for t in getattr(expr, "targets", []):
-                                targ = getattr(t, "target", t)
-                                if (
-                                    isinstance(targ, cst.Attribute)
-                                    and isinstance(getattr(targ, "value", None), cst.Name)
-                                    and getattr(getattr(targ, "value", None), "value", None) in ("self", "cls")
-                                ):
-                                    return True
-                    return False
-
-                # If multiple assignments occurred in setUp, prefer binding to a
-                # local name so cleanup rewrites are consistent and literal-only
-                # yields are avoided. Also, if the module already defines the
-                # conventional local base name (e.g., `_x_value`), force binding
-                # to avoid colliding with module-level identifiers.
-                # Force binding if module defines the conventional local name
-                # or if the module contains private/underscore-prefixed names
-                # (common pattern) which increases risk of collision with
-                # the conventional fixture-local names like `_x_value`.
-                force_bind_due_to_module_collision = base_local in module_level_names or any(
-                    name and name.startswith("_") for name in module_level_names
-                )
-                if (
-                    not multi_assigned
-                    and _is_literal(value_expr)
-                    and all(_is_simple_cleanup_statement(s) for s in spec.cleanup_statements)
-                    and not force_bind_due_to_module_collision
-                ):
-                    # yield the literal value and rewrite cleanup to use fixture name
-                    yield_stmt = cst.SimpleStatementLine(
-                        body=[cst.Expr(cst.Yield(cast(cst.BaseExpression, value_expr)))]
-                    )
-                    # accumulate as Any to accept mixed libcst statement flavors
-                    body_stmts_small_small: List[Any] = [yield_stmt]
-                    for stmt in spec.cleanup_statements:
-                        new_stmt = cast(Any, stmt).visit(_AttrRewriterTarget(attr, fname))
-                        body_stmts_small_small.append(new_stmt)
-                    # IndentedBlock expects Sequence[BaseStatement]; widen types here
-                    body = cst.IndentedBlock(body=list(cast(Sequence[cst.BaseStatement], body_stmts_small_small)))
-                    func = cst.FunctionDef(
-                        name=cst.Name(fname), params=cst.Parameters(), body=body, decorators=[decorator]
-                    )
-                    fixture_nodes.append(func)
-                else:
-                    # bind to local_name and yield it; rewrite cleanup to local_name
-                    # create assignment to local_name then yield it
-                    assign_stmt = cst.SimpleStatementLine(
-                        body=[
-                            cst.Assign(
-                                targets=[cst.AssignTarget(target=cst.Name(local_name))],
-                                value=cast(cst.BaseExpression, value_expr),
-                            )
-                        ]
-                    )
-                    yield_stmt = cst.SimpleStatementLine(body=[cst.Expr(cst.Yield(cst.Name(local_name)))])
-                    # accumulate as Any to accept Assign (BaseStatement) and
-                    # SimpleStatementLine/other small-statement flavors
-                    body_stmts_small_small = [assign_stmt, yield_stmt]
-                    for stmt in spec.cleanup_statements:
-                        new_stmt = cast(Any, stmt).visit(_AttrRewriterTarget(attr, local_name))
-                        body_stmts_small_small.append(new_stmt)
-                    body = cst.IndentedBlock(body=list(cast(Sequence[cst.BaseStatement], body_stmts_small_small)))
-                    func = cst.FunctionDef(
-                        name=cst.Name(fname), params=cst.Parameters(), body=body, decorators=[decorator]
-                    )
-                    fixture_nodes.append(func)
-            else:
-                # For simple literal values, return the literal directly instead
-                # of binding to a local name, preserving the original intent
-                # (e.g., return 42). For ambiguous self-referential expressions,
-                # delegate to guarded simple fixture creation to avoid broken
-                # placeholders like `str(schema_file)`.
-                # Treat simple scalars and literal containers/comprehensions as
-                # candidates for emitting a concrete return annotation.
-                is_container_literal = isinstance(value_expr, (cst.List, cst.Tuple, cst.Set, cst.Dict))
-                is_comp = getattr(cst, "ListComp", None) and isinstance(value_expr, getattr(cst, "ListComp"))
-                if _is_literal(value_expr) or is_container_literal or is_comp:
-                    # infer return annotation for the literal container/value
-                    # infer return annotation for the literal container/value
-                    ann: Optional[cst.Annotation]
-                    ann = None
-                    try:
-                        ann_res, names_req = _infer_ann(
-                            value_expr if isinstance(value_expr, cst.BaseExpression) else None
-                        )
-                        ann = ann_res
-                        for n in names_req:
-                            all_typing_needed.add(n)
-                    except Exception:
-                        ann = None
-                    return_stmt = cst.SimpleStatementLine(body=[cst.Return(cast(cst.BaseExpression, value_expr))])
-                    body = cst.IndentedBlock(body=[return_stmt])
-                    if ann is not None:
-                        func = cst.FunctionDef(
-                            name=cst.Name(fname),
-                            params=cst.Parameters(),
-                            returns=ann,
-                            body=body,
-                            decorators=[decorator],
-                        )
-                    else:
-                        func = cst.FunctionDef(
-                            name=cst.Name(fname), params=cst.Parameters(), body=body, decorators=[decorator]
-                        )
-                    fixture_nodes.append(func)
-                else:
-                    # Use guarded simple fixture creator which detects self-referential placeholders
-                    guarded = create_simple_fixture_with_guard(fname, cast(cst.BaseExpression, value_expr))
-                    fixture_nodes.append(guarded)
-
-            # Post-process the generated fixture node(s) to rewrite references
-            # to attributes included in a composite `temp_dirs` into accesses
-            # of the emitted `temp_dirs` mapping, and add parameters for
-            # dependencies on `temp_dirs` and other per-attribute fixtures.
-            try:
-                # Determine if this attr's value_expr references any handled attrs
-                needs_temp_dirs = False
-                other_deps: list[str] = []
-                ve = value_expr
-                # Prefer a rendered-code check for 'self.' occurrences to catch
-                # attributes embedded in complex literals (e.g., dicts). Fall
-                # back to structural _references_attribute when rendering fails.
-                try:
-                    rendered = (
-                        cst.Module(body=[cst.SimpleStatementLine(body=[cst.Expr(ve)])]).code if ve is not None else ""
-                    )
-                except Exception:
-                    rendered = ""
-                if "self." in rendered:
-                    # if any handled attr name appears in the rendered code, we
-                    # need to rewrite references to temp_dirs[...] in the
-                    # generated fixture body.
-                    for handled in handled_attrs:
-                        if f"self.{handled}" in rendered or f"{handled}" in rendered:
-                            needs_temp_dirs = True
-                            break
-                else:
-                    for handled in handled_attrs:
-                        if _references_attribute(ve, handled):
-                            needs_temp_dirs = True
-                            break
-                # detect references to other attributes which will be fixtures
-                for other_attr in cls.setup_assignments.keys():
-                    if other_attr == attr or other_attr in handled_attrs:
-                        continue
-                    if _references_attribute(ve, other_attr):
-                        other_deps.append(other_attr)
-
-                # build transformer to replace self.<handled> -> temp_dirs["handled"]
-                class _ReplaceSelfWithTempDirs(cst.CSTTransformer):
+                # replace self.* with bare names
+                class _ReplaceSelf(cst.CSTTransformer):
                     def leave_Attribute(self, original: cst.Attribute, updated: cst.Attribute) -> cst.BaseExpression:
                         if isinstance(original.value, cst.Name) and original.value.value in ("self", "cls"):
                             if isinstance(original.attr, cst.Name):
-                                name = original.attr.value
-                                if name in handled_attrs:
-                                    return cst.Subscript(
-                                        value=cst.Name("temp_dirs"),
-                                        slice=[
-                                            cst.SubscriptElement(slice=cst.Index(value=cst.SimpleString(repr(name))))
-                                        ],
-                                    )
+                                return cst.Name(original.attr.value)
                         return updated
 
-                # prepare params list
-                params: list[cst.Param] = []
-                if needs_temp_dirs:
-                    # annotate temp_dirs: Dict[str, Path]
-                    all_typing_needed.update({"Dict", "Path", "Any"})
-                    dict_ann = cst.Annotation(
-                        annotation=cst.Subscript(
-                            value=cst.Name("Dict"),
-                            slice=[
-                                cst.SubscriptElement(slice=cst.Index(value=cst.Name("str"))),
-                                cst.SubscriptElement(slice=cst.Index(value=cst.Name("Path"))),
-                            ],
-                        )
-                    )
-                    params.append(cst.Param(name=cst.Name("temp_dirs"), annotation=dict_ann))
+                call_in_fixture = assigned_call.visit(_ReplaceSelf())
+                targets = [cst.Name(local) for local, _, _ in sorted_group]
+                assign_target = cst.Assign(
+                    targets=[cst.AssignTarget(target=cst.Tuple(elements=[cst.Element(value=t) for t in targets]))],
+                    value=call_in_fixture,
+                )
+                assign_stmt = cst.SimpleStatementLine(body=[assign_target])
 
-                # add other deps as parameters (no annotation unless inferred)
-                for d in other_deps:
-                    # attempt to infer annotation from earlier spec if present
-                    ann_param = None
-                    if d in specs:
-                        dep_spec = specs[d]
-                        try:
-                            ann_node, names_req = (
-                                _infer_ann(dep_spec.value_expr)
-                                if isinstance(dep_spec.value_expr, cst.BaseExpression)
-                                else _infer_ann(None)
-                            )
-                            ann_param = ann_node
-                            for n in names_req:
-                                all_typing_needed.add(n)
-                        except Exception:
-                            ann_param = None
-                    params.append(cst.Param(name=cst.Name(d), annotation=ann_param))
+                ctor_args = [cst.Arg(value=cst.Name(local)) for local, _, _ in sorted_group]
+                ctor = cst.Call(func=cst.Name(namedtuple_name), args=ctor_args)
+                yield_stmt = cst.SimpleStatementLine(body=[cst.Expr(cst.Yield(ctor))])
+                decorator = cst.Decorator(decorator=cst.Attribute(value=cst.Name("pytest"), attr=cst.Name("fixture")))
+                body = cst.IndentedBlock(body=[assign_stmt, yield_stmt])
+                fixture_func = cst.FunctionDef(
+                    name=cst.Name(fixture_name), params=cst.Parameters(), body=body, decorators=[decorator]
+                )
 
-                # apply replacements on all fixture nodes we just added for this attr
-                for i in range(len(fixture_nodes) - 1, -1, -1):
-                    node = fixture_nodes[i]
-                    if isinstance(node, cst.FunctionDef) and node.name.value == fname:
-                        new_node = node
-                        # rewrite self.<handled> -> temp_dirs[...] if needed
-                        if needs_temp_dirs and ve is not None:
-                            new_body = new_node.body.visit(_ReplaceSelfWithTempDirs())
-                            new_node = new_node.with_changes(body=new_body)
-                        # attach params if any
-                        if params:
-                            new_node = new_node.with_changes(params=cst.Parameters(params=params))
-                        # special-case main_config annotation: prefer Dict[str, Any]
-                        if fname == "main_config":
-                            all_typing_needed.update({"Dict", "Any"})
-                            ann = cst.Annotation(
-                                annotation=cst.Subscript(
-                                    value=cst.Name("Dict"),
-                                    slice=[
-                                        cst.SubscriptElement(slice=cst.Index(value=cst.Name("str"))),
-                                        cst.SubscriptElement(slice=cst.Index(value=cst.Name("Any"))),
-                                    ],
-                                )
-                            )
-                            new_node = new_node.with_changes(returns=ann)
-                        # Attach transformed node back into fixture_nodes.
-                        fixture_nodes[i] = cast(cst.BaseStatement, new_node)
-                        break
-            except Exception:
-                # be conservative: skip post-processing on unexpected shapes
-                pass
-    result: dict[str, Any] = {"fixture_specs": specs, "fixture_nodes": fixture_nodes}
-    # No hard-coded snippet replacements here; rely on systematic
-    # generation and transformers to produce canonical fixtures.
-    if all_typing_needed:
-        result["needs_typing_names"] = sorted(all_typing_needed)
-    if used_shutil:
+                # prepend class_def and fixture to fixture_nodes
+                fixture_nodes.insert(0, class_def)
+                fixture_nodes.insert(1, fixture_func)
+                # request typing import names
+                try:
+                    needs_typing.update({"NamedTuple", "Any"})
+                except Exception:
+                    pass
+    except Exception:
+        pass
+
+    result: dict[str, Any] = {"fixture_specs": fixture_specs, "fixture_nodes": fixture_nodes}
+    if needs_typing:
+        result["needs_typing_names"] = sorted(needs_typing)
+    if needs_shutil:
         result["needs_shutil_import"] = True
     return result
+
+
+# Backwards-compatibility helpers for tests that import the module and
+# expect `generator_stage` and `_is_literal` at module level. These are
+# lightweight aliases that map to the canonical implementation above.
+def generator_stage(context: dict[str, Any]) -> dict[str, Any]:
+    return generator(context)
+
+
+def _is_literal(node: Any) -> bool:
+    """Compatibility: checks whether a node is a literal scalar.
+
+    Tests historically called generator._is_literal; keep a thin wrapper
+    around the internal scalar check used in this module.
+    """
+    try:
+        return isinstance(node, (cst.SimpleString, cst.Integer, cst.Float)) or (
+            isinstance(node, cst.Name) and getattr(node, "value", None) in ("True", "False")
+        )
+    except Exception:
+        return False
