+++ splurge_unittest_to_pytest/stages/pipeline.py
@@ -7,10 +7,7 @@ from typing import Any
 import libcst as cst
 from .collector import Collector
 
-try:
-    from .generator_v2 import generator_v2 as generator_stage
-except Exception:
-    from .generator import generator_stage
+from .generator import generator as generator_stage
 from .fixture_injector import fixture_injector_stage
 from .rewriter import rewriter_stage
 from .import_injector import import_injector_stage
