+++ tests/goldens/generator_migrated_composite.golden
@@ -1,3 +1,21 @@
+```plaintext
+import tempfile
+import shutil
+from pathlib import Path
+import pytest
+
+
+@pytest.fixture
+def temp_dir():
+    _temp_dir_value = tempfile.mkdtemp()
+    yield _temp_dir_value
+    shutil.rmtree(_temp_dir_value, ignore_errors=True)
+
+
+@pytest.fixture
+def config_dir(temp_dir):
+    return Path(temp_dir) / "conf"
+```
 @pytest.fixture
 def temp_dir():
     return "/tmp/foo"
