+++ tests/integration/test_generator_integration.py
@@ -2,7 +2,7 @@ import libcst as cst
 from libcst import MetadataWrapper
 
 from splurge_unittest_to_pytest.stages.collector import Collector
-from splurge_unittest_to_pytest.stages import generator
+from splurge_unittest_to_pytest.stages.generator import generator
 
 
 def test_generator_creates_fixture_from_collector_and_handles_collision():
@@ -18,7 +18,7 @@ def test_generator_creates_fixture_from_collector_and_handles_collision():
     out = coll.as_output()
 
     ctx = {"collector_output": out, "module": module}
-    res = generator.generator_stage(ctx)
+    res = generator(ctx)
 
     specs = res.get("fixture_specs")
     nodes = res.get("fixture_nodes")
@@ -27,6 +27,9 @@ def test_generator_creates_fixture_from_collector_and_handles_collision():
     spec = specs["a"]
     # Should detect cleanup and mark yield_style True
     assert spec.yield_style is True
-    # local_value_name should be set and different from conventional base if collision
-    assert spec.local_value_name is not None
-    assert nodes and any(n.name.value == "a" for n in nodes)
+    # When module already defines a colliding name, generator should bind to
+    # a local whose name includes the conventional base `_a_value`.
+    # Verify this by rendering fixture nodes and searching for the conventional
+    # local name fragment.
+    rendered_all = "\n\n".join(cst.Module(body=[n]).code for n in nodes)
+    assert "_a_value" in rendered_all
