+++ tests/unit/test_generator_bundle_namedtuple.py
@@ -1,7 +1,7 @@
 import libcst as cst
 from libcst import MetadataWrapper
 from splurge_unittest_to_pytest.stages.collector import Collector
-from splurge_unittest_to_pytest.stages.generator_v2 import generator_v2 as generator_stage
+from splurge_unittest_to_pytest.stages.generator import generator as generator_stage
 
 UNIT = '''
 class TestInitAPI(unittest.TestCase):
