+++ tests/unit/test_generator_nested_inference.py
@@ -1,5 +1,5 @@
 import libcst as cst
-from splurge_unittest_to_pytest.stages.generator_v2 import generator_v2 as generator_stage
+from splurge_unittest_to_pytest.stages.generator import generator as generator_stage
 from splurge_unittest_to_pytest.stages.collector import CollectorOutput, ClassInfo
