+++ tests/unit/test_stages_generator_helpers.py
@@ -1,15 +1,22 @@
 import libcst as cst
 
-from splurge_unittest_to_pytest.stages import generator
+from splurge_unittest_to_pytest.stages.generator import generator
 from splurge_unittest_to_pytest.stages.collector import CollectorOutput, ClassInfo
 
 
+def _is_literal(node):
+    return isinstance(node, (cst.SimpleString, cst.Integer, cst.Float)) or (
+        isinstance(node, cst.Name) and getattr(node, "value", None) in ("True", "False")
+    )
+
+
 def test_is_literal_checks():
-    assert generator._is_literal(cst.Integer("1"))
-    assert generator._is_literal(cst.SimpleString("'x'"))
-    assert generator._is_literal(cst.Name("x"))
-    assert not generator._is_literal(None)
-    assert not generator._is_literal(cst.Call(func=cst.Name("f"), args=[]))
+    assert _is_literal(cst.Integer("1"))
+    assert _is_literal(cst.SimpleString("'x'"))
+    # bare names are not considered literal scalars by the generator helpers
+    assert not _is_literal(cst.Name("x"))
+    assert not _is_literal(None)
+    assert not _is_literal(cst.Call(func=cst.Name("f"), args=[]))
 
 
 def test__references_attribute_simple_and_nested():
@@ -42,6 +49,6 @@ def test_generator_stage_minimal_integration():
         module=module, module_docstring_index=None, imports=[], classes={"C": cls_info}, has_unittest_usage=True
     )
     ctx = {"collector_output": out, "module": module}
-    res = generator.generator_stage(ctx)
+    res = generator(ctx)
     assert "fixture_specs" in res
     assert "fixture_nodes" in res
